% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{as.data.table.long.table}
\alias{as.data.table.long.table}
\title{Coerce a LongTable into a \code{data.table}}
\usage{
\method{as.data.table}{long.table}(x)
}
\arguments{
\item{x}{\code{LongTable} to coerce to a \code{data.table}}
}
\value{
A \code{data.table} containing the data from the LongTable, as well
as the `longTableDataMapper' attribute which contains the data needed to
reverse the coercion.
}
\description{
S3 version of coerce method for convenience.
}
