% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{as}
\alias{as}
\title{LongTable to data.table conversion}
\arguments{
\item{to}{\code{character} Class name to coerce to, currently only 'data.table'
and 'data.frame' are supported}

\item{from}{\code{SummarizedExperiment} object.}
}
\value{
A \code{data.table} with the data from a LongTable.

\code{data.table} containing the data from the LongTable, with the
`longTableDataMapper' attribute containg the metadata needed to reverse
the coercing operation.

\code{LongTable} object configured with the longTableDataMapper

\code{data.table} with long format of data in \code{from}

\code{data.frame} with long format of data in \code{from}.
}
\description{
Coerce a LongTable into a \code{data.table}.

Currently only supports coercing to data.table or data.frame

Coerce a data.table with the proper configuration attributes
back to a LongTable
}
\examples{
as(merckLongTable, 'data.table')

dataTable <- as(merckLongTable, 'data.table')
print(attr(dataTable, 'LongTableDataMapper')) # Method doesn't work without this
as(dataTable, 'LongTable')

SE <- molecularProfilesSlot(clevelandSmall_cSet)[[1]]
as(SE, 'data.table')

SE <- molecularProfilesSlot(clevelandSmall_cSet)[[1]]
as(SE, 'data.frame')

}
