% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTableDataMapper-class.R
\name{assay,LongTableDataMapper,ANY-method}
\alias{assay,LongTableDataMapper,ANY-method}
\title{Extract the data for an assay from a \code{LongTableDataMapper}}
\usage{
\S4method{assay}{LongTableDataMapper,ANY}(x, i, withDimnames = TRUE)
}
\arguments{
\item{x}{\code{LongTableDataMapper} The object to retrive assay data form according
to the \code{assayMap} slot.}

\item{i}{\code{character(1)} Name of an assay in the \code{assayMap} slot of \code{x}.}

\item{withDimNames}{\code{logical(1)} For compatibility with
\code{SummarizedExperiment::assay} generic. Not used.}
}
\value{
\code{data.table} Data for the specified assay extracted from the
\code{rawdata} slot of \code{x}.
}
\description{
Extract the data for an assay from a \code{LongTableDataMapper}
}
\keyword{internal}
