% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREDataMapper-class.R
\name{assay,TREDataMapper,ANY-method}
\alias{assay,TREDataMapper,ANY-method}
\title{Extract the data for an assay from a \code{TREDataMapper}}
\usage{
\S4method{assay}{TREDataMapper,ANY}(x, i, withDimnames = TRUE)
}
\arguments{
\item{x}{\code{TREDataMapper} The object to retrive assay data form according
to the \code{assayMap} slot.}

\item{i}{\code{character(1)} Name of an assay in the \code{assayMap} slot of \code{x}.}

\item{withDimnames}{\code{logical(1)} For compatibility with
\code{SummarizedExperiment::assay} generic. Not used.}
}
\value{
\code{data.table} Data for the specified assay extracted from the
\code{rawdata} slot of \code{x}.
}
\description{
Extract the data for an assay from a \code{TREDataMapper}
}
\keyword{internal}
