% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-class.R
\name{assayIndex,LongTable-method}
\alias{assayIndex,LongTable-method}
\title{Retrieve a copy of the assayIndex from the \verb{@.intern} slot.}
\usage{
\S4method{assayIndex}{LongTable}(x)
}
\arguments{
\item{`x`}{A \code{LongTable} or inheriting class.}
}
\value{
A \code{mutable} copy of the "assayIndex" for \code{x}
}
\description{
Retrieve a copy of the assayIndex from the \verb{@.intern} slot.
}
