% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{assayIndex}
\alias{assayIndex}
\title{Retrieve and assayIndex}
\usage{
assayIndex(x, ...)
}
\arguments{
\item{x}{An \code{S4} object.}

\item{...}{\code{pairlist} Allow definition of new parameters for
implementations of this generic.}
}
\value{
An object representing the "assayIndex" of an \code{S4} object.
}
\description{
Retrieve and assayIndex
}
\examples{
print("Generics shouldn't need examples?")

}
