% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-class.R
\name{assayKeys,LongTable-method}
\alias{assayKeys,LongTable-method}
\title{Retrieve a copy of the assayKey from the \verb{@.intern} slot.}
\usage{
\S4method{assayKeys}{LongTable}(x, i)
}
\arguments{
\item{`x`}{A \code{LongTable} or inheriting class.}

\item{`i`}{An optional valid assay name or index in \code{x}.}
}
\value{
A \code{mutable} copy of the "assayIndex" for \code{x}
}
\description{
Retrieve a copy of the assayKey from the \verb{@.intern} slot.
}
