% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{assayKeys}
\alias{assayKeys}
\title{Retrieve a set of assayKeys}
\usage{
assayKeys(x, ...)
}
\arguments{
\item{x}{An \code{S4} object.}

\item{...}{\code{pairlist} Allow definition of new parameters for
implementations of this generic.}
}
\value{
An object representing the "assayKeys" of an \code{S4} object.
}
\description{
Retrieve a set of assayKeys
}
\examples{
print("Generics shouldn't need examples?")

}
