% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{assayMap,LongTableDataMapper-method}
\alias{assayMap,LongTableDataMapper-method}
\title{assayMap}
\usage{
\S4method{assayMap}{LongTableDataMapper}(object)
}
\arguments{
\item{object}{An \code{LongTableDataMapper} object to access the rowDataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
A \code{list} of character vectors. The name of each list item will be
the assay in a \code{LongTableDataMapper} object that the columns in the
\code{character} vector will be assigned to.
}
\description{
assayMap
}
