% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{assayMap<-}
\alias{assayMap<-}
\title{assayMap<-}
\usage{
assayMap(object, ...) <- value
}
\arguments{
\item{object}{A \code{DataMapper}}

\item{...}{Allow new arguments to be defined on this generic.}

\item{value}{A \code{list} of character vectors, where the name of each list
item is the name of an assay and the values of each character vector
specify the columns mapping to the assay in the \code{S4} object the
\code{DataMapper} constructs.}
}
\value{
\code{invisible} Updates object in-place and returns and invisible NULL.
}
\description{
assayMap<-
}
