% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R,
%   R/LongTableDataMapper-class.R
\name{assayMap}
\alias{assayMap}
\title{assayMap}
\usage{
assayMap(object, ...)

assayMap(object, ...)
}
\arguments{
\item{object}{An \code{LongTableDataMapper} object to access the assayMap from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
A \code{list} of character vectors. The name of each list item will be
the assay in an \code{DataMapper} object that the columns in the \code{character}
vector will be assigned to.

A \code{list} of character vectors. The name of each list item will be
the assay in an \code{DataMapper} object that the columns in the \code{character}
vector will be assigned to.
}
\description{
assayMap

assayMap
}
