% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTableDataMapper-class.R
\name{assays,LongTableDataMapper-method}
\alias{assays,LongTableDataMapper-method}
\title{Extract the data for all assays from a \code{LongTableDataMapper}}
\usage{
\S4method{assays}{LongTableDataMapper}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{\code{LongTableDataMapper} The object to retrive assay data form according
to the \code{assayMap} slot.}

\item{withDimNames}{\code{logical(1)} For compatibility with
\code{SummarizedExperiment::assay} generic. Not used.}
}
\value{
\code{list} Data for all assays extracted from the
\code{rawdata} slot of \code{x} as a \code{list} of \code{data.tables}, where the \code{keys} for
each table are their \code{id_columns}.
}
\description{
Extract the data for all assays from a \code{LongTableDataMapper}
}
\keyword{internal}
