% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildComboProfiles}
\alias{buildComboProfiles}
\title{Build an assay table with an \code{S4} object.}
\usage{
buildComboProfiles(object, ...)
}
\arguments{
\item{object}{\code{S4} An S4 object a list-like slot containing assays for the
object.}

\item{...}{Allow new arguments to be defined for this generic.}
}
\value{
\code{data.table}.
}
\description{
Build an assay table with an \code{S4} object.
}
\examples{
"This is a generic method!"

}
