% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{buildLongTable,character-method}
\alias{buildLongTable,character-method}
\title{LongTable build method from character}
\usage{
\S4method{buildLongTable}{character}(from, rowDataCols, colDataCols, assayCols)
}
\arguments{
\item{from}{\code{character} Path to the .csv file containing the data and
metadata from which to build the \code{LongTable}.}

\item{rowDataCols}{\code{list} List with two \code{character} vectors, the first
specifying one or more columns to be used as cell identifiers (e.g.,
cell-line name columns) and the second containing any additional metadata
columns related to the cell identifiers.}

\item{colDataCols}{\code{list} List with two \code{character} vectors, the first
specifying one or more columns to be used as column identifiers (e.g.,
drug name columns) and the second containing any additional metadata
columns related to the column identifiers.}

\item{assayCols}{\code{list} A named list of character vectors specifying how to
parse assay columns into a list of \code{data.table}s. Each list data.table
will be named for the name of corresponding list item and contain the columns
specified in the character vector of column names in each list item.}
}
\value{
A \code{LongTable} object containing one or more assays, indexed by
rowID and colID.
}
\description{
LongTable Create a LongTable object from a single .csv file
}
