% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{buildLongTable}
\alias{buildLongTable}
\title{Build a LongTable object}
\usage{
buildLongTable(from, ...)
}
\arguments{
\item{from}{What to build the LongTable from?}

\item{...}{\code{pairlist} Allow definition of new parameters for
implementations of this generic.}
}
\value{
Depends on the implemented method
}
\description{
Build a LongTable object
}
\examples{
print("Generics shouldn't need examples?")

}
