% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immutable-class.R
\name{c.immutable}
\alias{c.immutable}
\title{Intercept concatenation for "immutable" class objects to return
another "immutable" class object.}
\usage{
\method{c}{immutable}(x, ...)
}
\arguments{
\item{x}{An R object inheriting from the "immutable" S3-clas}

\item{...}{Objects to concatenate to \code{x}.}
}
\description{
Ensures that \code{c} and \code{append} to an "immutable" class object return an
immutable class object.
}
