% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-utils.R
\name{$<-,LongTable-method}
\alias{$<-,LongTable-method}
\title{Update an assay from a LongTable object}
\usage{
\S4method{$}{LongTable}(x, name) <- value
}
\arguments{
\item{x}{A \code{LongTable} to update an assay for.}

\item{name}{\code{character(1)} The name of the assay to update}

\item{value}{A \code{data.frame} or \code{data.table} to update the assay with.}
}
\value{
Updates the assay \code{name} in \code{x} with \code{value}, returning an invsible
NULL.
}
\description{
Update an assay from a LongTable object
}
\examples{
merckLongTable$sensitivity <- merckLongTable$sensitivity

}
