% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cellInfo.R
\name{cellInfo<-}
\alias{cellInfo<-}
\title{cellInfo<- Generic}
\usage{
cellInfo(object) <- value
}
\arguments{
\item{object}{The \code{CoreSet} to replace cell info in}

\item{value}{A \code{data.frame} with the new cell annotations}
}
\value{
Updated \code{CoreSet}
}
\description{
Generic for cellInfo replace method
}
\examples{
cellInfo(clevelandSmall_cSet) <- cellInfo(clevelandSmall_cSet)

}
