% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-accessors.R
\name{cellInfo}
\alias{cellInfo}
\title{dateCreated<- Generic}
\usage{
cellInfo(object, ...)
}
\arguments{
\item{object}{A \code{CoreSet}}

\item{...}{Fallthrough arguements for defining new methods}

\item{value}{A \code{datetime} object to update the cSet with}
}
\value{
The date the CoreSet was created
}
\description{
A generic for the dateCreated method
}
\examples{
dateCreated(clevelandSmall_cSet) <- date()

}
