% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellNames-methods.R
\name{cellNames<-}
\alias{cellNames<-}
\title{cellNames<- Generic}
\usage{
cellNames(object, ...) <- value
}
\arguments{
\item{object}{The \code{CoreSet} to update}

\item{...}{Fallthrough arguments for defining new methods}

\item{value}{A \code{character} vector of the new cell names}
}
\value{
Updated \code{CoreSet}
}
\description{
A generic for the cellNames replacement method
}
\examples{
cellNames(clevelandSmall_cSet) <- cellNames(clevelandSmall_cSet)

}
