% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cellNames.R
\name{cellNames}
\alias{cellNames}
\title{cellNames Generic}
\usage{
cellNames(object, ...)
}
\arguments{
\item{object}{The \code{CoreSet} to return cell names from}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
A vector of the cell names used in the CoreSet
}
\description{
A generic for the cellNames method
}
\examples{
cellNames(clevelandSmall_cSet)

}
