% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clevelandSmall_cSet}
\alias{clevelandSmall_cSet}
\title{Cleaveland_mut RadioSet subsetted and cast as CoreSet}
\format{
CoreSet object
}
\usage{
data(clevelandSmall_cSet)
}
\description{
This dataset is just a dummy object derived from the Cleveland_mut RadioSet
  in the RadioGx R package. It's contents should not be interpreted and it
  is only present to test the functions in this package and provide
  examples
}
\references{
Lamb et al. The Connectivity Map: using gene-expression signatures to connect small molecules, genes, and disease. Science, 2006.
}
\keyword{datasets}
