% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTableDataMapper-accessors.R,
%   R/TREDataMapper-class.R
\name{colData,LongTableDataMapper-method}
\alias{colData,LongTableDataMapper-method}
\title{Convenience method to subset the \code{colData} out of the \code{rawdata} slot using
the assigned \code{colDataMap} metadata.}
\usage{
\S4method{colData}{LongTableDataMapper}(x, key = TRUE)

\S4method{colData}{LongTableDataMapper}(x, key = TRUE)
}
\arguments{
\item{x}{\code{TREDataMapper} object with valid data in the \code{rawdata} and
\code{colDataMap} slots.}

\item{key}{\code{logical(1)} Should the table be keyed according to the
\code{id_columns} of the \code{colDataMap} slot? This will sort the table in memory.
Default is TRUE.}
}
\value{
\code{data.table} The \code{colData} as specified in the \code{colDataMap} slot.

\code{data.table} The \code{colData} as specified in the \code{colDataMap} slot.
}
\description{
Convenience method to subset the \code{colData} out of the \code{rawdata} slot using
the assigned \code{colDataMap} metadata.

Convenience method to subset the \code{colData} out of the \code{rawdata} slot using
the assigned \code{colDataMap} metadata.
}
