% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R,
%   R/LongTableDataMapper-class.R
\name{colDataMap,LongTableDataMapper-method}
\alias{colDataMap,LongTableDataMapper-method}
\title{colDataMap}
\usage{
\S4method{colDataMap}{LongTableDataMapper}(object)

\S4method{colDataMap}{LongTableDataMapper}(object)
}
\arguments{
\item{object}{An \code{S4} object to access the rowDataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
A \code{list} with two items, the first being all columns needed to
uniquely identify each column in a \code{LongTable} and the second any
additional column-level metadata columns.

A \code{list} with two items, the first being all columns needed to
uniquely identify each column in a \code{LongTable} and the second any
additional column-level metadata columns.
}
\description{
colDataMap

colDataMap
}
