% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R,
%   R/LongTableDataMapper-class.R
\name{colDataMap<-,LongTableDataMapper,list_or_List-method}
\alias{colDataMap<-,LongTableDataMapper,list_or_List-method}
\title{colDataMap<-}
\usage{
\S4method{colDataMap}{LongTableDataMapper,list_or_List}(object) <- value

\S4method{colDataMap}{LongTableDataMapper,list_or_List}(object) <- value
}
\arguments{
\item{object}{A \code{LongTableDataMapper} object to assign a colDataMap to.}

\item{value}{A \code{list} or \code{List} where the first item is the named of the
identifier columns -- columns needed to uniquely identify each row in
colData -- and the second item is the metadata associated with those
the identifier columns, but not required to uniquely identify rows in
the object colData.}
}
\value{
None, modifies the object.

None, modifies the object.
}
\description{
colDataMap<-

colDataMap<-
}
