% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R,
%   R/LongTableDataMapper-class.R
\name{colDataMap}
\alias{colDataMap}
\title{colDataMap}
\usage{
colDataMap(object, ...)

colDataMap(object, ...)
}
\arguments{
\item{object}{An \code{S4} object to access the colDataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
A \code{list} with two items, the first being all columns needed to
uniquely identify each column in a \code{S4} object and the second any
additional column-level metadata columns.

A \code{list} with two items, the first being all columns needed to
uniquely identify each column in a \code{S4} object and the second any
additional column-level metadata columns.
}
\description{
colDataMap

colDataMap
}
