% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{colIDs}
\alias{colIDs}
\title{Generic to access the row identifiers for an object.}
\usage{
colIDs(object, ...)
}
\arguments{
\item{object}{\code{S4} An object to get column id columns from.}

\item{...}{ALlow new arguments to this generic}
}
\value{
Depends on the implemented method.
}
\description{
Generic to access the row identifiers for an object.
}
\examples{
print("Generics shouldn't need examples?")

}
