% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{comboResponse}
\alias{comboResponse}
\title{Grouping monotherapy responses with dual-therapy responses
within an assay.}
\usage{
comboResponse(assay_dt)
}
\arguments{
\item{assay_dt}{A \code{data.table} that must contain fields \code{drug1id}, \code{drug2id},
\code{drug1dose}, \code{drug2dose}, \code{cellid}, and \code{viability}.
The input assay data table contains both monotherapeutic observations
and dual-therapeutic observations.
Monotherapeutic observations have \code{NA} for either \code{drug1id} and \code{drug1dose},
or \code{drug2id} and \code{drug2dose}.}
}
\value{
A \code{data.table} containing fields
\code{drug1id}, \code{drug1dose}, \code{drug2id}, \code{drug2dose}, \code{cellid},
\code{viability}, \code{avg_viability_1}, and \code{avg_viability_2}.
Each \verb{avg_viability_*} is the response of a single-agent in the combination
averaged over its monotherapeutic replicates.
}
\description{
For preparing an input to synergy score method.
}
