% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{combotherapyPrefix}
\alias{combotherapyPrefix}
\title{Retrieve the configured combotherapy prefix}
\usage{
combotherapyPrefix()
}
\value{
\code{character(1)} The configured pattern for sample identifiers in CoreGx.
}
\description{
Retrieve the configured combotherapy prefix
}
\examples{
combotherapyPrefix()

}
