% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{CoreGx-deprecated}
\alias{CoreGx-deprecated}
\alias{CoreGx-defunct}
\title{List of \code{deprecated} or \code{defunct} methods in the \code{CoreGx} R package.}
\description{
List of \code{deprecated} or \code{defunct} methods in the \code{CoreGx} R package.
}
\details{
\subsection{deprecated}{

\code{CoreSet}: The \code{CoreSet} constructor is being updated to have a new API. This
API is currently available via the \code{CoreSet2} constructor. In Bioconductor
3.16, the old constructor will be renamed \code{CoreSet2} and the new constructor
will be renamed \code{CoreSet}.
}

\subsection{defunct}{

\code{buildLongTable}: This function no longer works as building a \code{LongTable} or
\code{TreatmentResponseExperiment} now uses a \code{DataMapper} and the \code{metaConstruct}
method. See \code{vignette("LongTable")} for a detailed description of how
to create a LongTable object.
}
}
