% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{CoreGx-options}
\alias{CoreGx-options}
\title{CoreGx Package Options}
\description{
A set of customizable options for the default behavior of CoreGx. Used
to set the default identifiers for various entities worked with
by the CoreGx package, including samples, treatments, and
tissues and assays.
}
\details{
\subsection{Configurable Identifiers}{
\subsection{Entity Identifiers}{

This set of identifiers is used to match to standardized columns used to
uniquely identify different entities within the object. Strings containing
"<n>" will have a numeric value inserted which corresponds to the index of
that identifier if for example there are multiple treatments or samples
for each observation in a given experiment. When only one of each
identifier is present in the experiment, the numeric portion will be
excluded. If your configured entity identifier doesn't contain "<n>" then
the numeric index will be appended to the end of the string.
\itemize{
\item sampleid: A string which contains the standardized identifier for samples
in the experiment. Defaults to "sample<n>id".
\item treatmentid: A string which contains the standardized
identifer for treatments in the experiment. Defaults to "treatment<n>id".
\item treatmentdose: A string which contains the standardized
identifier for treatment doses in the experiment. Defaults to
"treatment<n>dose".
\item tissueid: A string which contains the standardized
identifier for tissue type of samples in the experiment. Defaults to
"tissue<n>id".
}
}

\subsection{Assay Identifers}{

The assay identifiers are used to establish naming conventions for the
different kinds of data stored within a \verb{CoreSet`` and }TreatmentResponseExperiment`. They are a set of prefixed and suffixes
which are combined together so that CoreGx understands where in the object
to look for specific kinds of measurements needed to compute various
dose-response and drug synergy metrics. Prefixes and suffixes are combined
with an "\emph{" to create the assay names. For example, monotherapy viablity
measurement would be stored in an assay called: "<monotherapy>}<viability>",
where the values inside angle brackets are replaced with the configured
string for each prefix and suffix. It is recommended to leave these default
unless you are an advanced user of R and CoreGx.
\itemize{
\item raw: A string representing the standardized suffix for the raw data
assay in the experiment. Defaults to "raw". When no prefix is present
then the suffix will be used alone (e.g., for all your raw data the assay
would simply be named "raw").
\item viability: A string which contains the standardized suffix for viability
measurement assays in the experiment. Viability assays contain the
normalized treatment response values for samples in an experiment.
\item profiles: A string representing the standardized suffix for profile
assays in the experiment. Profiles are summary metrics computed over
a viability assay and can include things like dose-response curve parameters
or curve summary metrics such as IC50, AAC, etc.
\item monotherapy: A string representing the prefix for monotherapy assays
in the experiment. Defaults to "mono".
\item combotherapy: A string representing the prefix for combination therapy
assays in the experiment. Defaults to "combo".
}
}

}
}
