% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-class.R, R/CoreSet-accessors.R,
%   R/methods-molecularProfilesSlot.R, R/methods-name.R, R/methods-pertNumber.R,
%   R/methods-phenoInfo.R, R/methods-sensNumber.R
\docType{class}
\name{CoreSet-class}
\alias{CoreSet-class}
\alias{.CoreSet}
\alias{sensitivitySlot,CoreSet-method}
\alias{sensitivitySlot<-,CoreSet,list-method}
\alias{molecularProfilesSlot,CoreSet-method}
\alias{molecularProfilesSlot<-,CoreSet,list-method}
\alias{name,CoreSet-method}
\alias{name<-,CoreSet-method}
\alias{pertNumber,CoreSet-method}
\alias{pertNumber<-,CoreSet,array-method}
\alias{phenoInfo,CoreSet-method}
\alias{phenoInfo<-,CoreSet,character,data.frame-method}
\alias{phenoInfo<-,CoreSet,character,DataFrame-method}
\alias{sensNumber,CoreSet-method}
\alias{sensNumber<-,CoreSet,matrix-method}
\title{CoreSet - A generic data container for molecular profiles and
treatment response data}
\usage{
\S4method{sensitivitySlot}{CoreSet}(object)

\S4method{sensitivitySlot}{CoreSet,list}(object) <- value

\S4method{molecularProfilesSlot}{CoreSet}(object)

\S4method{molecularProfilesSlot}{CoreSet,list}(object) <- value

\S4method{name}{CoreSet}(object)

\S4method{name}{CoreSet}(object) <- value

\S4method{pertNumber}{CoreSet}(object)

\S4method{pertNumber}{CoreSet,array}(object) <- value

\S4method{phenoInfo}{CoreSet}(object, mDataType)

\S4method{phenoInfo}{CoreSet,character,data.frame}(object, mDataType) <- value

\S4method{phenoInfo}{CoreSet,character,DataFrame}(object, mDataType) <- value

\S4method{sensNumber}{CoreSet}(object)

\S4method{sensNumber}{CoreSet,matrix}(object) <- value
}
\arguments{
\item{object}{A \code{CoreSet} to extract the raw sensitivity data from}

\item{value}{A \code{list} of new sensitivity slot data for the rSet}
}
\description{
CoreSet - A generic data container for molecular profiles and
treatment response data
}
\details{
The CoreSet (CSet) class was developed as a superclass for pSets in the
PharmacoGx and RadioGx packages to contain the data generated in screens
of cancer cell lines for their genetic profile and sensitivities to therapy
(Pharmacological or Radiation). This class is meant to be a superclass which
is contained within the PharmacoSet (pSet) and RadioSet (RSet) objects
exported by PharmacoGx and RadioGx. The format of the data is similar for
both pSets and rSets, allowing much of the code to be abstracted into
the CoreSet super-class. However, the models involved with quantifying
cellular response to Pharmacological and Radiation therapy are widely
different, and extension of the cSet class allows the
packages to apply the correct model for the given data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{sensitivitySlot}: Retrieve the contents of the sensitivity slot

\item \code{sensitivitySlot<-}: Set the raw dose and viability data for a cSet and return
and updated copty

\item \code{molecularProfilesSlot}: Return a list containing all molecularProfiles in the cSet

\item \code{molecularProfilesSlot<-}: Update the contents of the molecularProfiles slot in a
CoreSet and returns an update copy

\item \code{name}: Return the name of the CoreSet

\item \code{name<-}: Return the name of the CoreSet

\item \code{pertNumber}: Return the summary of available perturbation
experiments

\item \code{pertNumber<-}: Update the summary of available perturbation
experiments

\item \code{phenoInfo}: Return the experiment info from the given type of molecular data in CoreSet

\item \code{phenoInfo<-}: Update the given type of molecular data experiment info in the CoreSet

\item \code{phenoInfo<-}: Update the given type of molecular data experiment info in the CoreSet

\item \code{sensNumber}: Return the summary of available sensitivity
experiments

\item \code{sensNumber<-}: Update the summary of available sensitivity
experiments
}}

\section{Slots}{

\itemize{
\item annotation: A \code{list} of annotation data about the \code{CoreSet},
including the \verb{$name} and the session information for how the object
was created, detailing the exact versions of R and all the packages used.
\item molecularProfiles: A \code{list} or \code{MultiAssayExperiment} containing
\code{CoreSet} object.
\item cell: A \code{data.frame} containg the annotations for all the cell
lines profiled in the data set, across all molecular data types and
treatment response experiments.
\item sensitivity: A \code{list} or \code{LongTable} containing all the data for the
sensitivity experiments, including \verb{$info}, a \code{data.frame} containing the
experimental info, \verb{$raw} a 3D \code{array} containing raw data,
\verb{$profiles}, a \code{data.frame} containing sensitivity profiles
statistics, and \verb{$n}, a \code{data.frame} detailing the number of
experiments for each cell-drug/radiationInfo pair
\item perturbation: \code{list} containing \verb{$n}, a \code{data.frame}
summarizing the available perturbation data. This slot is currently
being deprecated.
\item curation: \code{list} containing mappings for
\code{cell}, \code{tissue} names used in the data set to universal
identifiers used between different \code{CoreSet} objects
\item datasetType: \code{character} string of 'sensitivity',
'perturbation', or both detailing what type of data can be found in the
CoreSet, for proper processing of the data
}
}

\seealso{
\code{\link{CoreSet-accessors}}
}
