% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-utils.R
\name{CoreSet-utils}
\alias{CoreSet-utils}
\alias{subsetBySample,CoreSet-method}
\alias{subsetBySample}
\title{Utility methods for a \code{CoreSet} object.}
\usage{
\S4method{subsetBySample}{CoreSet}(x, samples)
}
\arguments{
\item{x}{A \code{CoreSet} object.}

\item{samples}{A \code{character} vector of sample names. Must be valid rownames
from \code{cellInfo(x)}.}
}
\value{
See details.
}
\description{
Documentation for utility methods for a \code{CoreSet} object, such as
set operations like subset and intersect. See @details for information
on different types of methods and their implementations.
}
\details{
\subsection{subset methods}{

\strong{subsetBySample}: Subset a \code{CoreSet} object by sample identifier.
\itemize{
\item value: a \code{CoreSet} object containing only \code{samples}.
}
}
}
\examples{
data(clevelandSmall_cSet)

## subset methods

### subsetBySample
samples <- cellInfo(clevelandSmall_cSet)$cellid[seq_len(10)]
clevelandSmall_cSet_sub <- subsetBySample(clevelandSmall_cSet, samples)

}
