% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{cSetName}
\alias{cSetName}
\title{cSetName Generic}
\usage{
cSetName(cSet)
}
\arguments{
\item{cSet}{A \code{CoreSet}}
}
\value{
The name of the CoreSet
}
\description{
A generic for the cSetName method
}
\examples{
cSetName(Cleveland_small)

}
