% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{DataMapper-accessors}
\alias{DataMapper-accessors}
\alias{rawdata,DataMapper-method}
\alias{rawdata,LongTableDataMapper-method}
\alias{rawdata}
\title{Accessing and modifying data in a \code{LongTableDataMapper} object.}
\usage{
\S4method{rawdata}{DataMapper}(object)
}
\arguments{
\item{object}{A \code{LongTableDataMapper} object to get or set data from.}

\item{...}{See details.}
}
\value{
Accessors: See details

Setters: An update \code{LongTableDataMapper} object, returned invisibly.
}
\description{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{LongTableDataMapper} object.
}
\details{
\strong{rawdata}: Get the raw data slot from a \code{LongTableDataMapper} object. Returns
a list-like containing one or more raw data inputs to the
\code{LongTableDataMapper} object.
}
