% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\docType{class}
\name{DataMapper-class}
\alias{DataMapper-class}
\alias{.DataMapper}
\title{An S4 Class For Mapping from Raw Experimental Data to a Specific S4 Object}
\description{
This object will be used as a way to abstract away data preprocessing.
}
\section{Slots}{

\itemize{
\item rawdata: A list-like object containing one or more pieces of raw data
that will be processed and mapped to the slots of an \code{S4} object.
}
}

