% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{rawdata}
\alias{rawdata}
\alias{rawdata,DataMapper-method}
\title{A generic function for extracting raw data from an \code{S4} object.}
\usage{
rawdata(object, ...)

\S4method{rawdata}{DataMapper}(object)
}
\arguments{
\item{object}{A \code{DataMapper} object to extract raw data from.}
}
\value{
The raw data object.

A list-like containing one or more raw data inputs to the
\code{DataMapper} object.
}
\description{
A generic function for extracting raw data from an \code{S4} object.

Get the raw data slot from a \code{DataMapper} object.
}
\section{Functions}{
\itemize{
\item \code{rawdata}: 

\item \code{rawdata,DataMapper-method}: 
}}

