% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateCreated-methods.R
\name{dateCreated<-}
\alias{dateCreated<-}
\title{dateCreated<- Generic}
\usage{
dateCreated(object, ...) <- value
}
\arguments{
\item{object}{A \code{CoreSet}}

\item{...}{Fallthrough arguements for defining new methods}

\item{value}{A \code{datetime} object to update the cSet with}
}
\value{
The date the CoreSet was created
}
\description{
A generic for the dateCreated method
}
\examples{
dateCreated(clevelandSmall_cSet) <- date()

}
