% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\docType{methods}
\name{dim,CoreSet-method}
\alias{dim,CoreSet-method}
\title{Get the dimensions of a CoreSet}
\usage{
\S4method{dim}{CoreSet}(x)
}
\arguments{
\item{x}{CoreSet}
}
\value{
A named vector with the number of Cells and Drugs in the CoreSet
}
\description{
Get the dimensions of a CoreSet
}
