% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linePtDist.R
\name{distancePointLine}
\alias{distancePointLine}
\title{Calculate shortest distance between point and line}
\usage{
distancePointLine(x, y, a, b, c)
}
\arguments{
\item{x}{x-coordinate of point}

\item{y}{y-coordinate of point}

\item{a}{coefficient in line equation a * x + b * y + c = 0}

\item{b}{coefficient in line equation a * x + b * y + c = 0}

\item{c}{coefficient in line equation a * x + b * y + c = 0}
}
\description{
This function calculates the shortest distance between a point and a line in 2D space.
}
\examples{
distancePointLine(0, 0, 1, -1, 1)

}
