% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-subset.R
\name{.}
\alias{.}
\title{Convenience function for converting R code to a call}
\usage{
.(...)
}
\arguments{
\item{...}{[`parilist`] One or more R expressions to convert to calls.}
}
\value{
[`call`] An R call object containing the quoted expression.
}
\description{
This is used to pass through unevaluated R expressions into subset and
  `[`, where they will be evaluated in the correct context.
}
\examples{
.(cell_line1 == 'A2058')

}
