% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.convertCSetMolecularProfilesToSE}
\alias{.convertCSetMolecularProfilesToSE}
\title{CSet molecularProfiles from ESets to SEs}
\usage{
.convertCSetMolecularProfilesToSE(cSet)
}
\arguments{
\item{cSet}{\code{S4} A CoreSet containing molecular data in ExpressionSets}
}
\value{
\code{S4} A CoreSet containing molecular data in a 
  SummarizedExperiments
}
\description{
Converts all ExpressionSet objects within the molecularProfiles slot of a 
  CoreSet to SummarizedExperiments
}
\keyword{internal}
