% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callingWaterfall.R
\name{.distancePointLine}
\alias{.distancePointLine}
\title{Calculate shortest distance between point and line}
\usage{
.distancePointLine(x, y, a = 1, b = 1, c = 0)
}
\arguments{
\item{x}{x-coordinate of point}

\item{y}{y-coordinate of point}

\item{a}{\code{numeric(1)} The coefficient in line equation a * x + b * y + c = 0.
Defaults to 1.}

\item{b}{\code{numeric(1)} The coefficient in line equation a * x + b * y + c = 0.
Defaults to 1.}

\item{c}{\code{numeric(1)} The intercept in line equation a * x + b * y + c = 0.
Defaults to 0.}
}
\value{
\code{numeric} The shortest distance between a point and a line.
}
\description{
This function calculates the shortest distance between a point
and a line in 2D space.
}
\examples{
.distancePointLine(0, 0, 1, -1, 1)

}
\keyword{internal}
