% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.fitCurve}
\alias{.fitCurve}
\title{.fitCurve}
\usage{
.fitCurve(
  x,
  y,
  f,
  density,
  step,
  precision,
  lower_bounds,
  upper_bounds,
  scale,
  family,
  median_n,
  trunc,
  verbose,
  gritty_guess,
  span = 1
)
}
\arguments{
\item{x}{[`numeric`] input/x values for function}

\item{y}{[`numeric`] output/y values for function}

\item{f}{[`function`] function f, parameterized by parameters to optimize}

\item{density}{[`numeric`] how many points in the dimension of each parameter should 
be evaluated (density of the grid)}

\item{step}{initial step size for pattern search.}

\item{precision}{[`numeric`] smallest step size used in pattern search, once step size drops below this value,
the search terminates.}

\item{lower_bounds}{[`numeric`] lower bounds for the paramater search space}

\item{upper_bounds}{[`numeric`] upper bounds for the parameter search space}

\item{scale}{[`numeric`] scale on which to measure probability for the error model (roughly SD of error)}

\item{family}{[`character`] which error family to use. Currently, `normal` and `cauchy` are implemented}

\item{median_n}{[`integer`] number of technical replicates per measured point in x. Used to 
evaluate the proper median distribution for the normal and cauchy error models}

\item{trunc}{[`logical`] Whether or not to truncate the values at 100% (1.0)}

\item{verbose}{[`logical`] should diagnostic messages be printed?}

\item{gritty_guess}{[`numeric`] intitial, uninformed guess on parameter values (usually heuristic)}

\item{span}{['numeric'] can be safely kept at 1, multiplicative ratio for initial step size in pattern search. 
Must be larger than precision.}
}
\description{
Curve optimization from 1 variable to 1 variable, using L-BFSG-B from optim, 
with fallback to pattern search if optimization fails to converge.
}
