% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTableDataMapper-class.R
\name{.get_dimData}
\alias{.get_dimData}
\title{Method to subset the rawdata with the corresponding dimensions "dimDataMap"
method.}
\usage{
.get_dimData(x, key, dim = c("row", "col", "meta"))
}
\arguments{
\item{x}{\code{LongTableDataMapper} or inheriting class.}

\item{key}{\code{logical(1)} Should the returned value be keyed by the 'id_columns'
item of the dimDataMap? Ignored when dim is "meta".}

\item{dim}{\code{character(1)} Which dimension should rawdata to subset for? Options
are "row", "col" and "meta", corresponding to the associated
slots of the \code{LongTableDataMapper}.}
}
\description{
Method to subset the rawdata with the corresponding dimensions "dimDataMap"
method.
}
\keyword{internal}
