% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTableDataMapper-accessors.R
\name{.get_dimDataFromMap}
\alias{.get_dimDataFromMap}
\title{Method to subset the rawdata with the corresponding dimensions "dimDataMap"
method.}
\usage{
.get_dimDataFromMap(x, key = TRUE, dataMap, funContext)
}
\arguments{
\item{x}{\code{LongTableDataMapper} or inheriting class.}

\item{key}{\code{logical(1)} Should the returned value be keyed by the 'id_columns'
item of the dimDataMap? Default is \code{TRUE}.}

\item{dataMap}{\code{list} The map of a \code{LongTableDataMapper} dimension, as
returned by the '*DataMap' methods.}

\item{funContext}{\code{character(1)} Contextual information about the calling
function, for debugging. Users don't need to worry about this.}
}
\description{
Method to subset the rawdata with the corresponding dimensions "dimDataMap"
method.
}
\keyword{internal}
