% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.getSupportVec}
\alias{.getSupportVec}
\title{.getSupportVec}
\usage{
.getSupportVec(x, output_length = 1001)
}
\arguments{
\item{x}{An input vector of dosages}

\item{output_length}{The length of the returned support vector}
}
\value{
\code{numeric} A numeric vector of interpolated concentrations
}
\description{
.getSupportVec
}
