% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{.longTableToSummarizedExperiment}
\alias{.longTableToSummarizedExperiment}
\title{Convert LongTable to gDR Style SummarizedExperiment}
\usage{
.longTableToSummarizedExperiment(LT, assay_names)
}
\arguments{
\item{LT}{\code{LongTable} to convert to gDR \code{SummarizedExperiment} format.}

\item{assay_names}{\code{character()} Names to rename the assays to. These
are assumed to be in the same order as \code{assayNames(LT)}.}
}
\value{
\code{SummarizedExperiment} object with all assay from \code{LT} as
\code{BumpyMatrix}es.
}
\description{
Convert LongTable to gDR Style SummarizedExperiment
}
