% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.meshEval}
\alias{.meshEval}
\title{meshEval}
\usage{
.meshEval(
  x,
  y,
  f,
  guess,
  lower_bounds,
  upper_bounds,
  density,
  n,
  scale,
  family,
  trunc
)
}
\arguments{
\item{x}{[`numeric`] input/x values for function}

\item{y}{[`numeric`] output/y values for function}

\item{f}{[`function`] function f, parameterized by parameters to optimize}

\item{guess}{[`numeric`] initial guess for the parameter values to improve upon}

\item{lower_bounds}{[`numeric`] lower bounds for the paramater search space}

\item{upper_bounds}{[`numeric`] upper bounds for the parameter search space}

\item{density}{[`numeric`] how many points in the dimension of each parameter should 
be evaluated (density of the grid)}

\item{n}{[`integer`] number of technical replicates per measured point in x. Used to 
evaluate the proper median distribution for the normal and cauchy error models}

\item{scale}{[`numeric`] scale on which to measure probability for the error model (roughly SD of error)}

\item{family}{[`character`] which error family to use. Currently, `normal` and `cauchy` are implemented}

\item{trunc}{[`logical`] Whether or not to truncate the values at 100% (1.0)}
}
\description{
generate an initial guess for dose-response curve parameters by evaluating
the residuals at different lattice points of the search space
}
