% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-updateS4.R
\name{.sensitivityToLongTable}
\alias{.sensitivityToLongTable}
\title{Convert the old sensitivity slot format into a \code{LongTable} and update the
\code{CoreSet} object.}
\usage{
.sensitivityToLongTable(object, mapper = FALSE)
}
\arguments{
\item{object}{Inheriting from \code{CoreSet}.}

\item{mapper}{Should the \code{LongTableDataMapper} object be early returned,
instead of the \code{LongTable} object. This can be useful if the conversion
fails or corrupts your data. You can then modify the \code{DataMapper} as
necessary to fix the sensititivity data.}
}
\value{
A \code{LongTable} constructed from \code{object@sensitivty}, or a
\code{LongTableDataMapper} if \code{mapper}=TRUE.
}
\description{
Convert the old sensitivity slot format into a \code{LongTable} and update the
\code{CoreSet} object.
}
\keyword{internal}
