% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-utils.R
\name{.subsetByIndex}
\alias{.subsetByIndex}
\title{Subset a \code{LongTable} using an "assayIndex" data.frame}
\usage{
.subsetByIndex(x, index, reindex = FALSE)
}
\arguments{
\item{x}{\code{LongTable}}

\item{index}{\code{data.frame} Table with columns "rowKey", "colKey" and
".<assayName>", were <assayName> is the value for each \code{assayNames(x)}.
Warning: rownames are dropped internally in coercion to \code{data.table},}

\item{reindex}{\code{logical(1)} Should index values be reset such that they
are the smallest possible set of consecutive integers. Modifies the
"rowKey", "colKey", and all assayKey columns.}
}
\value{
\code{LongTable} subset according to the provided index.
}
\description{
Subset a \code{LongTable} using an "assayIndex" data.frame
}
