% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-utils.R
\name{.subsetDimData}
\alias{.subsetDimData}
\title{Subset a \code{LongTable} object}
\usage{
.subsetDimData(x, i, dim = c("row", "col"), ...)
}
\arguments{
\item{x}{\code{LongTable}}

\item{i}{A valid query to the \code{i} parameter of the metadata \code{data.table} for
the selected dimension.}

\item{dim}{\code{character(1)} Either "row" or "col", depending if you want to
subset by the first or second table-like dimension of a \code{x}.}

\item{...}{Fallthrough arguments to the \verb{<dim>Data} function.}
}
\value{
\code{data.table} The dimData, subset with i.
}
\description{
Subset a \code{LongTable} object
}
