% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.table_is_subset}
\alias{.table_is_subset}
\title{Tests of table 1 is a subset of table2, in which case there will be no rows
is the set difference}
\usage{
.table_is_subset(table1, table2)
}
\description{
Tests of table 1 is a subset of table2, in which case there will be no rows
is the set difference
}
\keyword{internal}
