% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.verifyKeyIntegrity}
\alias{.verifyKeyIntegrity}
\title{Ensure that all rowID and colID keys are valid}
\usage{
.verifyKeyIntegrity(rowData, colData, assays)
}
\arguments{
\item{rowData}{A [`data.table`] containing row level annotations.}

\item{colData}{A [`data.table`] containing column level annotations for a
`LongTable`.}

\item{assays}{A [`list`] of `data.table`s, one for each assay in an `LongTable`.}
}
\description{
Ensure that all rowID and colID keys are valid
}
\keyword{internal}
