% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{drugInfo}
\alias{drugInfo}
\title{drugInfo Generic}
\usage{
drugInfo(cSet)
}
\arguments{
\item{cSet}{The \code{CoreSet} to retrieve drug info from}
}
\value{
a \code{data.frame} with the drug annotations
}
\description{
Generic for drugInfo method
}
\examples{
data(CCLEsmall)

}
