% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{drugNames<-}
\alias{drugNames<-}
\title{drugNames<- Generic}
\usage{
drugNames(object) <- value
}
\arguments{
\item{object}{The \code{CoreSet} to update}

\item{value}{A \code{character} vector of the new drug names}
}
\value{
Updated \code{CoreSet}
}
\description{
A generic for the drugNames replacement method
}
\examples{
data(CCLEsmall)
drugNames(CCLEsmall) <- drugNames(CCLEsmall)

}
