% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{drugNames}
\alias{drugNames}
\title{drugNames Generic}
\usage{
drugNames(cSet)
}
\arguments{
\item{cSet}{The \code{CoreSet} to return drug names from}
}
\value{
A vector of the drug names used in the CoreSet
}
\description{
A generic for the drugNames method
}
\examples{
data(CCLEsmall)
drugNames(CCLEsmall)

}
