% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{drugSensitivitySig}
\alias{drugSensitivitySig}
\title{Compute the correlation between a molecular feature and treatment response}
\usage{
drugSensitivitySig(object, ...)
}
\arguments{
\item{object}{An object inheriting form the `CoreGx::CoreSet` class}

\item{...}{Allow definition of new arguments to this generic}
}
\value{
A 3D array of genes x drugs x metric
}
\description{
Compute the correlation between a molecular feature and treatment response
}
