% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{endoaggregate}
\alias{endoaggregate}
\title{Perform aggregation over an S4 object, but return an object of the same
class.}
\usage{
endoaggregate(x, ...)
}
\arguments{
\item{x}{An \code{S4} method to endomorphically aggregate over.}

\item{...}{\code{pairlist} Allow definition of new parameters for
implementations of this generic.}
}
\value{
An object with the same class as \code{x}.
}
\description{
Perform aggregation over an S4 object, but return an object of the same
class.
}
\examples{
print("Generics shouldn't need examples?")

}
