% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examineGOF.R
\name{examineGOF}
\alias{examineGOF}
\title{Getter for attributes of an object}
\usage{
examineGOF(pars)
}
\arguments{
\item{pars}{The object for which attributes are to be returned}
}
\value{
A named vector where index `Rsquare` contains the attributes of the object
}
\description{
Getter for attributes of an object
}
