% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleDataMapper}
\alias{exampleDataMapper}
\title{Example LongTableDataMapper}
\format{
LongTableDataMapper object
}
\usage{
data(exampleDataMapper)
}
\description{
A dummy LongTableDataMapper object to be used in package examples.
}
\keyword{datasets}
