% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{featureInfo<-}
\alias{featureInfo<-}
\title{featureInfo<- Generic}
\usage{
featureInfo(object, mDataType) <- value
}
\arguments{
\item{object}{The \code{CoreSet} to replace gene annotations in}

\item{mDataType}{The type of molecular data to be updated}

\item{value}{A \code{data.frame} with the new feature annotations}
}
\value{
Updated \code{CoreSet}
}
\description{
Generic for featureInfo replace method
}
\examples{
featureInfo(Cleveland_small, "rna") <- featureInfo(Cleveland_small, "rna")

}
