% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{featureInfo}
\alias{featureInfo}
\title{featureInfo Generic}
\usage{
featureInfo(object, mDataType, ...)
}
\arguments{
\item{object}{The \code{CoreSet} to retrieve feature annotations from}

\item{mDataType}{the type of molecular data}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
a \code{data.frame} with the feature annotations
}
\description{
Generic for featureInfo method
}
\examples{
featureInfo(clevelandSmall, "rna")

}
