% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{fNames}
\alias{fNames}
\title{fNames Generic}
\usage{
fNames(cSet, mDataType)
}
\arguments{
\item{cSet}{The \code{CoreSet}}

\item{mDataType}{The molecular data type to return feature names for}
}
\value{
A \code{character} vector of the feature names
}
\description{
A generic for the fNames method
}
\examples{
fNames(Cleveland_small, "rna")

}
