% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getComboProfiles}
\alias{getComboProfiles}
\title{Get selected assay profiles for drug combinations}
\usage{
getComboProfiles(tre, profiles)
}
\arguments{
\item{tre}{\code{TreatmentResponseExperiment} contains curated drug combination data.}

\item{profiles}{\code{character} a vector of profile names, i.e., column names of assays.}
}
\value{
A \code{data.table} containing fields
\code{treatment1id}, \code{treatment1dose}, \code{treatment2id}, \code{treatment2dose}, \code{sampleid},
along with columns of selected profiles from their assays.
Each \verb{*_1} is the monothearpy profile of treatment 1 in the combination,
and the same rule applies to treatment2.
}
\description{
Get selected assay profiles for drug combinations
}
\examples{
\dontrun{
combo_profile_1 <- getComboProfiles(tre, c("auc", "SCORE"))
combo_profile_2 <- getComboProfiles(tre, c("HS", "EC50", "E_inf", "ZIP"))
}

}
