% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{getCoreGxOptions}
\alias{getCoreGxOptions}
\title{Retrieve all configured options for CoreGx.}
\usage{
getCoreGxOptions()
}
\value{
\code{list()} All configured options for the CoreGx pacakge.
}
\description{
Retrieve all configured options for CoreGx.
}
\examples{
getCoreGxOptions()

}
