% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-getIntern.R
\name{getIntern,LongTable,character-method}
\alias{getIntern,LongTable,character-method}
\title{Get the symbol(s) x from the object@.intern slot of a LongTable}
\usage{
\S4method{getIntern}{LongTable,character}(object, x, ...)
}
\arguments{
\item{object}{[`LongTable`]}

\item{x}{[`character`] One or more symbol name strings to retrieve from
the object@.intern environment.}

\item{...}{[`pairlist`] Addtional arguments to get or mget inside of the
function.}
}
\value{
value of x if length(x) == 1 else named list of values for all
    symbols in x
}
\description{
Get the symbol(s) x from the object@.intern slot of a LongTable
}
