% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-accessors.R
\name{getIntern<-,LongTable,immutable_list-method}
\alias{getIntern<-,LongTable,immutable_list-method}
\title{Set the .intern slot of a LongTable}
\usage{
\S4method{getIntern}{LongTable,immutable_list}(object) <- value
}
\arguments{
\item{object}{\code{LongTable}}

\item{value}{An \code{immutable_list} object, being a class union between \code{list}
and \code{immutable} S3 classes.}
}
\value{
Updates the object and returns invisibly.
}
\description{
Set the .intern slot of a LongTable
}
\keyword{internal}
