% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getIntern<-}
\alias{getIntern<-}
\title{Set the internal structural metadata for an S4 class}
\usage{
getIntern(object, ...) <- value
}
\arguments{
\item{object}{An R object to update internal structural metadata for.}

\item{value}{An \code{immutable_list} object, being a class union between \code{list}
and \code{immutable} S3 classes.}
}
\value{
Updates the object and returns invisibly.
}
\description{
Set the internal structural metadata for an S4 class
}
\examples{
print("Generics shouldn't need examples?")

}
\keyword{internal}
