% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getIntern}
\alias{getIntern}
\title{Retrieve the symbol for the object@.intern slot}
\usage{
getIntern(object, x, ...)
}
\arguments{
\item{object}{\code{S4} An object with an @.itern slot containing an environment.}

\item{x}{\code{character} One or more symbol names to retrieve from the
object@.intern environment.}

\item{...}{Allow new parmeters to be defined for this generic.}
}
\value{
Depends on the implemented method
}
\description{
Internal slot for storing metadata relevant to the internal operation of an
S4 object.
}
\details{
Warning: This method is intended for developer use and can be ignored by
users.
}
\examples{
print("Generics shouldn't need examples?")

}
