% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-guessMapping.R
\name{guessMapping}
\alias{guessMapping}
\title{Generic for Guessing the Mapping Between Some Raw Data and an S4 Object}
\usage{
guessMapping(object, ...)
}
\arguments{
\item{object}{An \code{S4} object containing so raw data to guess data to
object slot mappings for.}

\item{...}{Allow new arguments to be defined for this generic.}
}
\value{
An \code{S4} object with mapping guesses in the appropriate slots.
}
\description{
Generic for Guessing the Mapping Between Some Raw Data and an S4 Object
}
