% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-class.R
\name{idCols,LongTable-method}
\alias{idCols,LongTable-method}
\title{Retrieve the unique identifier columns used for primary keys in rowData and
colData.}
\usage{
\S4method{idCols}{LongTable}(object)
}
\arguments{
\item{object}{\code{LongTable}}
}
\value{
\code{character} A character vector containing the unique rowIDs and
colIDs in a LongTable object.
}
\description{
Retrieve the unique identifier columns used for primary keys in rowData and
colData.
}
\examples{
idCols(merckLongTable)

}
