% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-iteration.R
\name{lapply,MultiAssayExperiment-method}
\alias{lapply,MultiAssayExperiment-method}
\title{lapply
lapply method for \code{MultiAssayExperiment}}
\usage{
\S4method{lapply}{MultiAssayExperiment}(X, FUN, ...)
}
\arguments{
\item{X}{A \code{MultiAssayExperiment} object.}

\item{FUN}{A function to be applied to each \code{SummarizedExperiment} in a
in \code{X}.}

\item{...}{Fall through parameters to \code{FUN}}
}
\value{
A \code{MultiAssayExperiment} object, modifed such that
\code{experiments(X) <- endoapply(experiments(X), FUN, ...)}.s
}
\description{
lapply
lapply method for \code{MultiAssayExperiment}
}
