% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-class.R
\docType{class}
\name{list_or_LongTable-class}
\alias{list_or_LongTable-class}
\title{A class union to allow multiple types in a CoreSet slot}
\description{
A class union to allow multiple types in a CoreSet slot
}
