% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-accessors.R
\name{LongTable-accessors}
\alias{LongTable-accessors}
\title{Accessing and modifying information in a \code{LongTable}}
\value{
Accessors: See details.

Setters: An updated \code{LongTable} object, returned invisibly.
}
\description{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{LongTable} object.
}
\examples{
data(merckLongTable)
}
