% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DataMapper.R
\docType{class}
\name{LongTableDataMapper-class}
\alias{LongTableDataMapper-class}
\alias{.LongTableDataMapper}
\title{A Class for Mapping Between Raw Data and an `LongTable` Object}
\description{
This object will be used as a way to abstract away data preprocessing.
}
\section{Slots}{

\describe{
\item{\code{rowDataMap}}{A list of mappings from}

\item{\code{colDataMap}}{A list of mappings from}

\item{\code{assayMap}}{A list of mappings from}

\item{\code{metadataMap}}{A list of mappings from}

\item{\code{metadata}}{A list of mappings from}
}}

\keyword{internal}
