% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{LongTableDataMapper}
\alias{LongTableDataMapper}
\title{Constructor for the \code{LongTableDataMapper} class, which maps from one or
more raw experimental data files to the slots of a \code{LongTable} object.}
\usage{
LongTableDataMapper(
  rawdata,
  rowDataMap = list(),
  colDataMap = list(),
  assayMap = list(),
  metadataMap = list()
)
}
\arguments{
\item{rawdata}{A list-like}

\item{rowDataMap}{A list-like}

\item{colDataMap}{A list-like}

\item{assayMap}{A list-like}

\item{metadataMap}{A list-like}
}
\description{
Constructor for the \code{LongTableDataMapper} class, which maps from one or
more raw experimental data files to the slots of a \code{LongTable} object.
}
