% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\docType{methods}
\name{mDataNames,CoreSet-method}
\alias{mDataNames,CoreSet-method}
\title{mDataNames}
\usage{
\S4method{mDataNames}{CoreSet}(cSet)
}
\arguments{
\item{cSet}{PharamcoSet object}
}
\value{
Vector of names of the molecular data types
}
\description{
Returns the molecular data names for the CoreSet.
}
\examples{
data(CCLEsmall)
mDataNames(CCLEsmall)

}
