% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{mDataNames}
\alias{mDataNames}
\title{mDataNames}
\usage{
mDataNames(cSet)
}
\arguments{
\item{cSet}{PharamcoSet object}
}
\value{
Vector of names of the molecular data types
}
\description{
Returns the molecular data names for the CoreSet.
}
\examples{
data(CCLEsmall)
mDataNames(CCLEsmall)

}
