% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{merckLongTable}
\alias{merckLongTable}
\title{Merck Drug Combination Data LongTable}
\format{
LongTable object
}
\usage{
data(merckLongTable)
}
\description{
This is a LongTable object created from some drug combination data provided
to our lab by Merck.
}
\references{
TODO:: Include a reference
}
\keyword{datasets}
