% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAssays-method.R
\name{mergeAssays}
\alias{mergeAssays}
\title{Merge assays with an \code{S4} object.}
\usage{
mergeAssays(object, ...)
}
\arguments{
\item{object}{\code{S4} An S4 object a list-like slot containing assays for the
object.}

\item{...}{Allow new arguments to be defined for this generic.}
}
\value{
A modified version of \code{object}.
}
\description{
Merge assays with an \code{S4} object.
}
\examples{
"This is a generic method!"

}
