% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metadata.R
\name{metadata,LongTable-method}
\alias{metadata,LongTable-method}
\title{Getter method for the metadata slot of a \code{LongTable} object}
\usage{
\S4method{metadata}{LongTable}(x)
}
\arguments{
\item{x}{The \code{LongTable} object from which to retrieve the metadata list.}
}
\value{
\code{list} The contents of the \code{metadata} slot of the \code{LongTable}
object.
}
\description{
Getter method for the metadata slot of a \code{LongTable} object
}
