% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-accessors.R
\name{metadata<-,LongTable-method}
\alias{metadata<-,LongTable-method}
\title{Setter method for the metadata slot of a \code{LongTable} object}
\usage{
\S4method{metadata}{LongTable}(x) <- value
}
\arguments{
\item{x}{\code{LongTable} The LongTable to update}

\item{value}{\code{list} A list of new metadata associated with a \code{LongTable}
object.}
}
\value{
\code{LongTable} A copy of the \code{LongTable} object with the \code{value} in
the metadata slot.
}
\description{
Setter method for the metadata slot of a \code{LongTable} object
}
