% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{metadataMap,LongTableDataMapper-method}
\alias{metadataMap,LongTableDataMapper-method}
\title{metadataMap}
\usage{
\S4method{metadataMap}{LongTableDataMapper}(object)
}
\arguments{
\item{object}{A \code{LongTableDataMapper} object to access the rowDataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
A \code{list} of \code{character} vectors. Each item is an element of the
constructed objects \verb{@metadata} slot
}
\description{
metadataMap
}
