% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R,
%   R/LongTableDataMapper-class.R
\name{metadataMap<-,LongTableDataMapper,list_or_List-method}
\alias{metadataMap<-,LongTableDataMapper,list_or_List-method}
\title{metadataMap<-}
\usage{
\S4method{metadataMap}{LongTableDataMapper,list_or_List}(object) <- value

\S4method{metadataMap}{LongTableDataMapper,list_or_List}(object) <- value
}
\arguments{
\item{object}{A \code{DataMapper} object.}

\item{value}{A \code{list} of \code{character} vectors. The name of each list item
is the name of}

\item{...}{Allow new parameters for this generic.}
}
\value{
\code{invisible} Updates object in-place, then returns and
\code{invisible(NULL)}.

\code{invisible} Updates object in-place, then returns and
\code{invisible(NULL)}.
}
\description{
metadataMap<-

metadataMap<-
}
