% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{metadataMap<-}
\alias{metadataMap<-}
\title{metadataMap<-}
\usage{
metadataMap(object, ...) <- value
}
\arguments{
\item{object}{A \code{DataMapper} object.}

\item{...}{Allow new parameters for this generic.}

\item{value}{A \code{list} of \code{character} vectors. The name of each list item
is the name of the item in the \verb{@metadata} slot of the \code{S4} object created
when \code{metaConstruct} is called on the \code{DataMapper}, and the character
vector specifies the columns of \verb{@rawdata} to assign the item.}
}
\value{
\code{invisible} Updates object in-place, then returns and
\code{invisible(NULL)}.
}
\description{
metadataMap<-
}
