% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{metadataMap}
\alias{metadataMap}
\title{metadataMap}
\usage{
metadataMap(object, ...)
}
\arguments{
\item{object}{An \code{S4} object to access the metadataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
A \code{list} where each the name of each item
}
\description{
metadataMap
}
