% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-molecularProfiles.R
\name{molecularProfiles}
\alias{molecularProfiles}
\title{molecularProfiles Generic}
\usage{
molecularProfiles(object, mDataType, assay, ...)
}
\arguments{
\item{object}{The \code{CoreSet} to retrieve molecular profiles from}

\item{mDataType}{\code{character} The type of molecular data}

\item{assay}{\code{character} Name of the desired assay; if excluded
defaults to first assay in the SummarizedExperiment for the given
mDataType. Use \code{assayNames(molecularProfiles(object, dataType)}
to check which assays are available for a given molecular datatype.}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
a \code{matrix} of data for the given mDataType and assay
}
\description{
Generic for molecularProfiles method
}
\examples{
data(clevelandSmall_cSet)
molecularProfiles(clevelandSmall_cSet, "rna")

}
