% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{molecularProfilesList}
\alias{molecularProfilesList}
\title{molecularProfilesList Generic}
\usage{
molecularProfilesList(object, ...)
}
\arguments{
\item{object}{A \code{CoreSet} from which to return a list of all availble
SummarizedExperiment objects}
}
\value{
A \code{list} containing the molecularProfiles from a cSet

Generic for molecularProfilesList
}
\description{
molecularProfilesList Generic
}
