% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularProfilesSlot-methods.R
\name{molecularProfilesSlot}
\alias{molecularProfilesSlot}
\title{molecularProfilesSlot Generic}
\usage{
molecularProfilesSlot(object, ...)
}
\arguments{
\item{object}{A \code{CoreSet} from which to return a list of all availble
SummarizedExperiment objects}

\item{...}{A \code{list} of additional parameters; included to allow adding
arguments to methods on this generc}
}
\value{
A \code{list} containing the molecularProfiles from a cSet

Generic for molecularProfilesSlot
}
\description{
molecularProfilesSlot Generic
}
\examples{
data(clevelandSmall_cSet)
molecularProfilesSlot(clevelandSmall_cSet)

}
