% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{monotherapyPrefix}
\alias{monotherapyPrefix}
\title{Retrieve the configured monotherapy prefix}
\usage{
monotherapyPrefix()
}
\value{
\code{character(1)} The configured pattern for sample identifiers in CoreGx.
}
\description{
Retrieve the configured monotherapy prefix
}
\examples{
monotherapyPrefix()

}
