% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immutable-class.R
\name{mutable}
\alias{mutable}
\title{Remove the "immutable" S3-class from an R object, allowing it to be
modified normally again.}
\usage{
mutable(object)
}
\arguments{
\item{object}{An R object inheriting from the "immutable" class.}
}
\value{
The \code{object} with the "immutable" class stripped from it.
}
\description{
Remove the "immutable" S3-class from an R object, allowing it to be
modified normally again.
}
\examples{
immut_list <- immutable(list())
mutable(immut_list)

}
