% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-name.R
\name{name<-}
\alias{name<-}
\title{name<- Generic}
\usage{
name(object, ...) <- value
}
\arguments{
\item{object}{A \code{CoreSet} object}

\item{...}{Fallthrough arguements for defining new methods}

\item{value}{A \code{character} string with the name to assign to the cSet}
}
\value{
The name of the CoreSet
}
\description{
A generic for the name<- method
}
\examples{
name(clevelandSmall_cSet) <- "Cleveland Small"

}
