% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R, R/methods-name.R
\name{name}
\alias{name}
\alias{as.data.frame.LongTable}
\title{Coerce a LongTable to a data.frame}
\usage{
\method{as.data.frame}{LongTable}(x, row.names, optional = TRUE, ...)

name(object, ...)
}
\arguments{
\item{x}{[`LongTable`] to coerce to `data.frame`.}

\item{row.names}{An optional [`character`] vector of rownames. We do not
recommend using this parameter, it is included for S3 method consistency
with `as.data.frame`.}

\item{optional}{[`logical`] Is it optional for row and column names to be
valid R names? If FALSE will use the make.names function to ensure the
row and column names are valid R names. Defaults to TRUE.}

\item{...}{Fallthrough arguements for defining new methods}

\item{object}{A \code{CoreSet}}
}
\value{
[`data.frame`] containing the data from the LongTable, with the
  `LongTable.config' attribute containg the metadata needed to reverse
  the coercion operation.

The name of the CoreSet
}
\description{
S3 version of coerce method fro convenience.

A generic for the name method
}
\examples{
name(clevelandSmall_cSet)

}
