% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{pertNumber}
\alias{pertNumber}
\title{pertNumber Generic}
\usage{
pertNumber(object, ...)
}
\arguments{
\item{object}{A \code{CoreSet}}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
A 3D \code{array} with the number of perturbation experiments per drug and cell line, and data type
}
\description{
A generic for the pertNumber method
}
\examples{
pertNumber(clevelandSmall_cSet)

}
