% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-accessors.R
\name{sensitivityRaw<-,CoreSet,array-method}
\alias{sensitivityRaw<-,CoreSet,array-method}
\title{sensitivityRaw<- Setter}
\usage{
\S4method{sensitivityRaw}{CoreSet,array}(object) <- value
}
\arguments{
\item{object}{A \code{PharmacoSet} to extract the raw sensitivity data from}

\item{value}{A \code{array} containing the raw dose and viability data for the
pSet}
}
\value{
A copy of the \code{PharmacoSet} containing the updated sensitivty data
}
\description{
sensitivityRaw<- Setter
}
\section{Functions}{
\itemize{
\item \code{sensitivityRaw<-,CoreSet,array-method}: 
}}

\examples{
data(CCLEsmall)
sensitivityRaw(CCLEsmall) <- sensitivityRaw(CCLEsmall)

}
