% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{phenoInfo<-}
\alias{phenoInfo<-}
\title{phenoInfo<- Generic}
\usage{
phenoInfo(object, mDataType) <- value
}
\arguments{
\item{object}{The \code{CoreSet} to retrieve molecular experiment annotations from}

\item{mDataType}{the type of molecular data}

\item{value}{a \code{data.frame} with the new experiment annotations}
}
\value{
The updated \code{CoreSet}
}
\description{
Generic for phenoInfo replace method
}
\examples{

phenoInfo(Cleveland_small, mDataType="rna") <- phenoInfo(Cleveland_small, mDataType="rna")

}
