% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-class.R
\name{printSlot}
\alias{printSlot}
\title{Helper function to print slot information}
\usage{
printSlot(slotName, slotData)
}
\arguments{
\item{slotName}{\code{character} The name of the slot to print.}

\item{slotData}{\code{data.table} The data to print.}
}
\description{
Helper function to print slot information
}
\keyword{internal}
