% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{profileSuffix}
\alias{profileSuffix}
\title{Retrieve the configured profile assay suffix}
\usage{
profileSuffix()
}
\value{
\code{character(1)} The configured pattern for sample identifiers in CoreGx.
}
\description{
Retrieve the configured profile assay suffix
}
\examples{
profileSuffix()

}
