% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-methods.R
\name{annotation<-,CoreSet,list-method}
\alias{annotation<-,CoreSet,list-method}
\title{annotation<- Slot Setter}
\usage{
\S4method{annotation}{CoreSet,list}(object) <- value
}
\arguments{
\item{object}{A \code{RadioSet}}

\item{value}{A \code{list} of annotations to add to the annotatiosn slot of
an rSet}
}
\value{
A copy of the \code{CoreSet} with the updated annotation slot
}
\description{
annotation<- Slot Setter
}
\section{Functions}{
\itemize{
\item \code{annotation<-,CoreSet,list-method}: Update the annotation slot of a tSet
}}

\examples{
data(clevelandSmall_cSet)
annotation(clevelandSmall_cSet) <- annotation(clevelandSmall_cSet)

}
