% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DataMapper.R
\name{rawdata,DataMapper-method}
\alias{rawdata,DataMapper-method}
\title{Get the raw data slot from a \code{DataMapper} object.}
\usage{
\S4method{rawdata}{DataMapper}(object)
}
\arguments{
\item{object}{A \code{DataMapper} object to extract raw data from.}
}
\value{
A list-like containing one or more raw data inputs to the
\code{DataMapper} object.
}
\description{
Get the raw data slot from a \code{DataMapper} object.
}
