% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DataMapper.R
\name{rawdata}
\alias{rawdata}
\title{A generic function for extracting raw data from an \code{S4} object.}
\usage{
rawdata(object, ...)
}
\arguments{
\item{object}{An \code{S4} object to extract rawdata from.}
}
\value{
The raw data object.
}
\description{
A generic function for extracting raw data from an \code{S4} object.
}
