% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREDataMapper-class.R
\name{rowData,TREDataMapper-method}
\alias{rowData,TREDataMapper-method}
\title{Convenience method to subset the \code{rowData} out of the \code{rawdata} slot using
the assigned \code{rowDataMap} metadata.}
\usage{
\S4method{rowData}{TREDataMapper}(x, key = TRUE)
}
\arguments{
\item{x}{\code{TREDataMapper} object with valid data in the \code{rawdata} and
\code{colDataMap} slots.}

\item{key}{\code{logical(1)} Should the table be keyed according to the
\code{id_columns} of the \code{rowDataMap} slot? This will sort the table in memory.
Default is TRUE.}
}
\value{
\code{data.table} The \code{rowData} as specified in the \code{rowDataMap} slot.
}
\description{
Convenience method to subset the \code{rowData} out of the \code{rawdata} slot using
the assigned \code{rowDataMap} metadata.
}
