% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{rowDataMap<-,LongTableDataMapper,list_or_List-method}
\alias{rowDataMap<-,LongTableDataMapper,list_or_List-method}
\title{rowDataMap<-}
\usage{
\S4method{rowDataMap}{LongTableDataMapper,list_or_List}(object) <- value
}
\arguments{
\item{object}{A \code{LongTableDataMapper} object to access the rowDataMap from.}

\item{value}{A \code{list} or \code{List} where the first item is the names of the
identifier columns -- columns needed to uniquely identify each row in
rowData -- and the second item is the metadata associated with those
the identifier columns, but not required to uniquely identify rows in
the object rowData.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
\code{invisible} Updates object and returns an invisible NULL.
}
\description{
rowDataMap<-
}
