% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{rowDataMap<-}
\alias{rowDataMap<-}
\title{rowDataMap<-}
\usage{
rowDataMap(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} object to access the colDataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}

\item{value}{A like-like object containing the values to assign to the
\code{rowDataMap} slot of the \code{S4} object.}
}
\value{
\code{invisible} Update object in-place and returns and invisible NULL.
}
\description{
rowDataMap<-
}
