% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{rowDataMap}
\alias{rowDataMap}
\title{rowDataMap}
\usage{
rowDataMap(object, ...)
}
\arguments{
\item{object}{An \code{S4} object to access the rowDataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
\code{list} of two \code{character} vectors, the first are the columns required
to uniquely identify each row of a \code{LongTable} and the second any additional
row-level metadata.
}
\description{
rowDataMap
}
