% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{rowMeta}
\alias{rowMeta}
\title{Generic to access the row identifiers from}
\usage{
rowMeta(object, ...)
}
\arguments{
\item{object}{\code{\link{S4}} An object to get row metadata columns from.}

\item{...}{Allow new arguments to this generic.}
}
\value{
Depends on the implemented method.
}
\description{
Generic to access the row identifiers from
}
\examples{
print("Generics shouldn't need examples?")

}
