% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{sampleIdentifier}
\alias{sampleIdentifier}
\title{Retrieve the configured sample identifier}
\usage{
sampleIdentifier()
}
\value{
\code{character(1)} The configured pattern for sample identifiers in CoreGx.
}
\description{
Retrieve the configured sample identifier
}
\examples{
sampleIdentifier()

}
