% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityInfo<-}
\alias{sensitivityInfo<-}
\title{sensitivityInfo<- Generic Method}
\usage{
sensitivityInfo(object, ...) <- value
}
\arguments{
\item{object}{An [`S4`] object to set treatment response experiment
annotations for.}

\item{...}{Allow new arguments to be defined for this generic.}

\item{value}{The new treatment response experiment annotations.}
}
\value{
Depends on the implemented method
}
\description{
Generic function to get the annotations for a treatment response experiment
  from an S4 class.
}
\examples{
print("Generics shouldn't need examples?")

}
