% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{sensitivityInfo}
\alias{sensitivityInfo}
\title{sensitivityInfo Generic}
\usage{
sensitivityInfo(object, ...)
}
\arguments{
\item{object}{The \code{CoreSet} to retrieve sensitivity experiment annotations from}

\item{...}{Fallthrough arguments for defining new methods}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for sensitivityInfo method
}
\examples{
sensitivityInfo(clevelandSmall)

}
