% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityMeasures<-}
\alias{sensitivityMeasures<-}
\title{sensitivityMeasures<- Generic}
\usage{
sensitivityMeasures(object, ...) <- value
}
\arguments{
\item{object}{An [`S4`] object to update.}

\item{...}{Allow new methods to be defined for this generic.}

\item{value}{A set of names for sensitivity measures to use to
update the object with.}
}
\description{
Set the names of the sensitivity summary metrics available in an S4
  object.
}
