% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{sensitivityMeasures}
\alias{sensitivityMeasures}
\title{sensitivityMeasures Generi}
\usage{
sensitivityMeasures(object, ...)
}
\arguments{
\item{object}{The \code{CoreSet}}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
A \code{character} vector of all the available sensitivity measures
}
\description{
A generic for the sensitivityMeasures  method
}
\examples{
sensitivityMeasures(clevelandSmall_cSet)

}
