% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityProfiles<-}
\alias{sensitivityProfiles<-}
\title{sensitivityProfiles<- Generic}
\usage{
sensitivityProfiles(object, ...) <- value
}
\arguments{
\item{object}{An [`S4`] object to update the sensitivity profile summaries
for.}

\item{...}{Fallthrough arguments for defining new methods}

\item{value}{An object with the new sensitivity profiles. If a
matrix object is passed in, converted to data.frame before assignment}
}
\value{
Updated \code{CoreSet}
}
\description{
A generic for the sensitivityProfiles replacement method
}
