% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityProfiles}
\alias{sensitivityProfiles}
\title{sensitivityProfiles Generic}
\usage{
sensitivityProfiles(object, ...)
}
\arguments{
\item{object}{The [`S4`] object to retrieve sensitivity profile summaries
from.}

\item{...}{[`pairlist`] Allow defining new arguments for this generic.}
}
\description{
A generic for sensitivityProfiles getter method
}
