% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityRaw<-}
\alias{sensitivityRaw<-}
\title{sensitivityRaw<- Generic}
\usage{
sensitivityRaw(object, ...) <- value
}
\arguments{
\item{object}{An \code{\link{S4}} object to extract the raw sensitivity data from.}

\item{...}{\code{\link{pairlist}} Allow new parameters to be defined for this generic.}

\item{value}{An object containing dose and viability metrics to update
the object with.}
}
\value{
Depends on the implemented method
}
\description{
Generic function to set the raw data array for a treatment response experiment
in an S4 class.
}
