% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityRaw}
\alias{sensitivityRaw}
\title{sensitivityRaw Generic Method}
\usage{
sensitivityRaw(object, ...)
}
\arguments{
\item{object}{An [`S4`] object to extract the raw sensitivity experiment
data from.}

\item{...}{[`pairlist`]  Allow new parameters to be defined for this generic.}
}
\description{
Generic function to get the raw data array for a treatment response experiment
  from an S4 class.
}
