% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-accessors.R
\name{sensitivitySlot<-}
\alias{sensitivitySlot<-}
\title{sensitivitySlot<- Replacement Generic}
\usage{
sensitivitySlot(object, ...) <- value
}
\arguments{
\item{object}{A \code{CoreSet} to extract the raw sensitivity data from}

\item{...}{A \code{list} to allow new parameters in specific methods}

\item{value}{A \code{list} of new sensitivity slot data for the rSet}
}
\value{
A copy of the \code{CoreSet} containing the updated sensitivty slot
}
\description{
sensitivitySlot<- Replacement Generic
}
\examples{
data(clevelandSmall_cSet)
sensitivitySlot(clevelandSmall_cSet) <- sensitivitySlot(clevelandSmall_cSet)

}
