% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\docType{methods}
\name{show,CoreSet-method}
\alias{show,CoreSet-method}
\title{Show a CoreSet}
\usage{
\S4method{show}{CoreSet}(object)
}
\arguments{
\item{object}{\code{CoreSet}}
}
\value{
Prints the CoreSet object to the output stream, and returns invisible NULL.
}
\description{
Show a CoreSet
}
\examples{
show(Cleveland_small)

}
