% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{show,LongTable-method}
\alias{show,LongTable-method}
\title{Show method for the LongTable class}
\usage{
\S4method{show}{LongTable}(object)
}
\arguments{
\item{object}{A \code{\link{LongTable}} object to print the results for.}
}
\value{
\code{\link{invisible}} Prints to console.
}
\description{
Show method for the LongTable class
}
\examples{
show(merckLongTable)

}
