% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{showSigAnnot}
\alias{showSigAnnot}
\title{Get the annotations for a \code{Signature} class object, as returned by
\code{drugSensitivitysig} or \code{radSensitivtySig} functions available in
\code{PharmacoGx} and \code{RadioGx}, respectively.}
\usage{
showSigAnnot(object, ...)
}
\arguments{
\item{object}{A \code{Signature} class object}

\item{...}{Allow definition of new arguments to this generic}
}
\value{
NULL Prints the signature annotations to console
}
\description{
Get the annotations for a \code{Signature} class object, as returned by
\code{drugSensitivitysig} or \code{radSensitivtySig} functions available in
\code{PharmacoGx} and \code{RadioGx}, respectively.
}
\examples{
print("Generics shouldn't need examples?")

}
