% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\docType{methods}
\name{[,CoreSet-method}
\alias{[,CoreSet-method}
\title{`[`}
\usage{
\S4method{[}{CoreSet}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{CSet}

\item{i}{Cell lines to keep in CSet}

\item{j}{Drugs to keep in CSet}

\item{...}{further arguments}

\item{drop}{A boolean flag of whether to drop single dimensions or not}
}
\value{
Returns the subsetted CSet
}
\description{
`[`
}
