% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-utils.R
\name{[[<-,LongTable,ANY,ANY-method}
\alias{[[<-,LongTable,ANY,ANY-method}
\title{\verb{[[<-} Method for LongTable Class}
\usage{
\S4method{[[}{LongTable,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{A \code{LongTable} to update.}

\item{i}{The name of the assay to update, must be in \code{assayNames(object)}.}

\item{value}{A \code{data.frame}}
}
\value{
A \code{LongTable} object with the assay \code{i} updated using \code{value}.
}
\description{
Just a wrapper around assay<- for convenience. See
\verb{?'assay<-,LongTable,character-method'.}
}
\examples{
merckLongTable[['sensitivity']] <- merckLongTable[['sensitivity']]

}
