% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-subsetTo.R
\name{subsetTo}
\alias{subsetTo}
\title{Subset a CoreSet object based on various parameters, such as cell lines, molecular features}
\usage{
subsetTo(object, ...)
}
\arguments{
\item{object}{An object inheriting from the \code{CoreGx::CoreSet} class}

\item{...}{Allow definition of new arguments to this generic}
}
\value{
A subsetted version of the original \code{object}
}
\description{
Subset a CoreSet object based on various parameters, such as cell lines, molecular features
}
\examples{
"Generics shouldn't need examples!"

}
\keyword{internal}
