% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summarizeMolecularProfiles}
\alias{summarizeMolecularProfiles}
\title{Summarize molecular profile data such that there is a single entry for each cell line/treatment combination}
\usage{
summarizeMolecularProfiles(object, ...)
}
\arguments{
\item{object}{An object inheriting form the `CoreGx::CoreSet` class}

\item{...}{Allow definition of new arguments to this generic}
}
\value{
[`SummarizedExperiment`] containing the summarized molecular profile data
}
\description{
Summarize molecular profile data such that there is a single entry for each cell line/treatment combination
}
\examples{
# Dummy examples to satisfy Bioc requiremnts
setMethod('summarizeMolecularProfiles', signature('CoreSet'), function(object) object)

}
\keyword{internal}
