% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{summarizeSensitivityProfiles}
\alias{summarizeSensitivityProfiles}
\title{Summarize across replicates for a sensitivity dose-response experiment}
\usage{
summarizeSensitivityProfiles(object, ...)
}
\arguments{
\item{object}{An [`S4`] object to summarize sensitivity profiles for.}

\item{...}{Allow definition of new arguments to this generic}
}
\description{
Summarize across replicates for a sensitivity dose-response experiment
}
