% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{tissueIdentifer}
\alias{tissueIdentifer}
\title{Retrieve the configued tissue identifer pattern}
\usage{
tissueIdentifer()
}
\value{
\code{character(1)} The configured pattern for sample identifiers in CoreGx.
}
\description{
Retrieve the configued tissue identifer pattern
}
\examples{
tissueIdentifier()

}
