% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{treatmentIdentifier}
\alias{treatmentIdentifier}
\title{Retrieve the configure treatment identifer pattern}
\usage{
treatmentIdentifier()
}
\value{
\code{character(1)} The configured pattern for sample identifiers in CoreGx.
}
\description{
Retrieve the configure treatment identifer pattern
}
\examples{
treatmentIdentifier()

}
