% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentResponseExperiment-class.R
\docType{class}
\name{TreatmentResponseExperiment-class}
\alias{TreatmentResponseExperiment-class}
\alias{.TreatmentResponseExperiment}
\title{TreatmentResponseExperiment class definition}
\value{
\code{TreatmentResponseExperiment} object containing the assay data from
a treatment response experiment
}
\description{
Define a private constructor method to be used to build a
\code{TreatmentResponseExperiment} object.
}
\section{Slots}{

\describe{
\item{\code{rowData}}{See Slots section.}

\item{\code{colData}}{See Slots section.}

\item{\code{assays}}{See Slots section.}

\item{\code{metadata}}{See Slots section.}

\item{\code{.intern}}{See Slots section.}
}}

\section{Slots}{

\itemize{
\item \emph{rowData}: A \code{data.table} containing the metadata associated with the
row dimension of a \code{TreatmentResponseExperiment}.
\item \emph{colData}: A \code{data.table} containing the metadata associated with the
column dimension of a \code{TreatmentResponseExperiment}.
\item \emph{assays}: A \code{list} of \code{data.table}s, one for each assay in a
\code{TreatmentResponseExperiment}.
\item \emph{metadata}: An optional \code{list} of additional metadata for a
\code{TreatmentResponseExperiment} which doesn't map to one of the dimensions.
\item \emph{.intern}: An \code{environment} that holds internal structural metadata
about a \code{TreatmentResponseExperiment} object, such as which columns are
required to key the object. An environment has been used to allow locking
items, which can prevent accidental modification of a property required
for the class to work.
}
}

\keyword{internal}
