% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateObject-methods.R
\name{updateObject,CoreSet-method}
\alias{updateObject,CoreSet-method}
\title{Update the \code{CoreSet} class after changes in it struture or API}
\usage{
\S4method{updateObject}{CoreSet}(object, verify = FALSE)
}
\arguments{
\item{object}{A \code{CoreSet} object to update the class structure for.}
}
\value{
\code{CoreSet} with update class structure.
}
\description{
Update the \code{CoreSet} class after changes in it struture or API
}
