% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-class.R
\name{updateSampleId}
\alias{updateSampleId}
\title{Update the sample ids in a cSet object}
\usage{
updateSampleId(object, new.ids = vector("character"))
}
\arguments{
\item{object}{The object for which the sample ids will be updated}

\item{new.ids}{The new ids to assign to the object}
}
\value{
\code{CoreSet} The modified CoreSet object
}
\description{
Update the sample ids in a cSet object
}
\examples{
updateSampleId(clevelandSmall_cSet, sampleNames(clevelandSmall_cSet))

}
\keyword{internal}
