% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-class.R
\name{updateTreatmentId}
\alias{updateTreatmentId}
\title{Update the treatment ids in a cSet object}
\usage{
updateTreatmentId(object, new.ids = vector("character"))
}
\arguments{
\item{object}{The object for which the treatment ids will be updated}

\item{new.ids}{The new ids to assign to the object}
}
\value{
\code{CoreSet} The modified CoreSet object
}
\description{
Update the treatment ids in a cSet object
}
\examples{
updateTreatmentId(clevelandSmall_cSet, treatmentNames(clevelandSmall_cSet))

}
\keyword{internal}
