% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateCsetMolecularProfilesToSEConversion.R
\name{validateCsetMolecularProfilesToSEConversion}
\alias{validateCsetMolecularProfilesToSEConversion}
\title{Validate cSet molecularProfiles Conversion}
\usage{
validateCsetMolecularProfilesToSEConversion(cSet_old, cSet_new)
}
\arguments{
\item{cSet}{\code{S4} a cSet containing molecularProfiles as ExpressionSets}
}
\value{
\code{message} Any slots which are not the same
}
\description{
Checks that all the information contained in an ExpressionSet molecularProfile 
  was successfully tranferred to the SummarizedExperiment molecularProfile
}
