% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-options.R
\name{viabilitySuffix}
\alias{viabilitySuffix}
\title{Retrieve the configured viability assay suffix}
\usage{
viabilitySuffix()
}
\value{
\code{character(1)} The configured pattern for sample identifiers in CoreGx.
}
\description{
Retrieve the configured viability assay suffix
}
\examples{
viabilitySuffix()

}
