% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{assayCols}
\alias{assayCols}
\title{Generic to access the assay columns of a rectangular object.}
\usage{
assayCols(object, ...)
}
\arguments{
\item{object}{\code{\link{S4}} An object to get assay ids from.}

\item{...}{Allow new arguments to this generic.}
}
\value{
Depends on the implemented method.
}
\description{
Generic to access the assay columns of a rectangular object.
}
\examples{
print("Generics shouldn't need examples?")

}
