% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellInfo-methods.R
\name{cellInfo}
\alias{cellInfo}
\title{cellInfo Getter}
\usage{
cellInfo(object, ...)
}
\arguments{
\item{object}{The \code{CoreSet} to retrieve cell info from}

\item{...}{\code{list} Fall through arguments to allow generic
to be defined with different parameters}
}
\value{
a \code{data.frame} with the cell annotations
}
\description{
Get cell line information from a PharmacoSet object
}
\examples{
data(clevelandSmall_cSet)
cellInf <- cellInfo(clevelandSmall_cSet)

}
