% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{cellInfo}
\alias{cellInfo}
\title{cellInfo Generic}
\usage{
cellInfo(cSet)
}
\arguments{
\item{cSet}{The \code{CoreSet} to retrieve cell info from}
}
\value{
a \code{data.frame} with the cell annotations
}
\description{
Generic for cellInfo method
}
\examples{
data(Cleveland_small)
cellInfo(Cleveland_small)

}
