% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{cellNames}
\alias{cellNames}
\title{cellNames Generic}
\usage{
cellNames(cSet)
}
\arguments{
\item{cSet}{The \code{CoreSet} to return cell names from}
}
\value{
A vector of the cell names used in the CoreSet
}
\description{
A generic for the cellNames method
}
\examples{
cellNames(Cleveland_small)

}
