% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R,
%   R/LongTableDataMapper-class.R
\name{colDataMap<-}
\alias{colDataMap<-}
\title{colDataMap<-}
\usage{
colDataMap(object, ...) <- value

colDataMap(object, ...) <- value
}
\arguments{
\item{object}{An \code{S4} object to access the colDataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}

\item{value}{A list-like object containing the values to assign to the
\code{colDataMap} slot of the \code{S4} object.}
}
\value{
None, modifies the object.

None, modifies the object.
}
\description{
colDataMap<-

colDataMap<-
}
