% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{colMeta}
\alias{colMeta}
\title{Generic to access the column identifiers for a rectangular object.}
\usage{
colMeta(object, ...)
}
\arguments{
\item{object}{[`S4`] An object to get column metadata columns from.}

\item{...}{ALlow new arguments to this generic}
}
\value{
Depends on impemented method.
}
\description{
Generic to access the column identifiers for a rectangular object.
}
\examples{
print("Generics shouldn't need examples?")

}
