% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{CoreGx-deprecated}
\alias{CoreGx-deprecated}
\alias{CoreGx-defunct}
\title{List of \code{deprecated} or \code{defunct} methods in the \code{CoreGx} R package.}
\description{
List of \code{deprecated} or \code{defunct} methods in the \code{CoreGx} R package.
}
\details{
\subsection{deprecated}{
}

\subsection{defunct}{

\code{buildLongTable}: This function no longer works as building a \code{LongTable} or
\code{TreatmentResponseExperiment} now uses a \code{DataMapper} and the \code{metaConstruct}
method. See \code{vignette("LongTable")} for a detailed description of how
to create a LongTable object.
}
}
