% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-accessors.R
\name{CoreSet-accessors}
\alias{CoreSet-accessors}
\alias{annotation,CoreSet-method}
\alias{annotation}
\alias{annotation<-,CoreSet,list-method}
\alias{annotation<-}
\alias{dateCreated,CoreSet-method}
\alias{dateCreated}
\alias{dateCreated<-,CoreSet,character-method}
\alias{dateCreated<-,CoreSet-method}
\alias{dateCreated<-}
\alias{cellInfo,CoreSet-method}
\alias{cellInfo}
\alias{cellInfo<-,CoreSet,data.frame-method}
\alias{cellInfo<-}
\alias{cellNames,CoreSet-method}
\alias{cellName,CoreSet-method}
\alias{cellName}
\alias{cellNames<-,CoreSet,character-method}
\alias{cellNames<-,CoreSet,list-method}
\alias{cellNames<-}
\alias{curation,CoreSet-method}
\alias{curation}
\alias{curation<-,CoreSet,list-method}
\alias{curation<-}
\alias{datasetType,CoreSet-method}
\alias{datasetType}
\alias{datasetType<-,CoreSet,character-method}
\title{Accessing and modifying information in a \code{CoreSet}}
\usage{
\S4method{annotation}{CoreSet}(object)

\S4method{annotation}{CoreSet,list}(object) <- value

\S4method{dateCreated}{CoreSet}(object)

\S4method{dateCreated}{CoreSet,character}(object) <- value

\S4method{cellInfo}{CoreSet}(object)

\S4method{cellInfo}{CoreSet,data.frame}(object) <- value

\S4method{cellNames}{CoreSet}(object)

\S4method{cellNames}{CoreSet,character}(object) <- value

\S4method{curation}{CoreSet}(object)

\S4method{curation}{CoreSet,list}(object) <- value

\S4method{datasetType}{CoreSet}(object)

\S4method{datasetType}{CoreSet,character}(object) <- value
}
\arguments{
\item{object}{A \code{CoreSet} object.}

\item{value}{See details.}

\item{...}{See details.}
}
\value{
Accessors: See details.

Setters: An updated \code{CoreSet} object, returned invisibly.
}
\description{
Accessing and modifying information in a \code{CoreSet}
}
\details{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{CoreSet} object:

\strong{annotation}: A \code{list} of \code{CoreSet} annotations with items: 'name',
the name of the object; dateCreated, date the object was created; 'sessionInfo',
the \code{sessionInfo()} when the object was created; 'call', the R constructor call;
and 'version', the object version.

\strong{annotation<-}: Setter method for the annotation slot. Arguments:
\itemize{
\item value: a \code{list} of annotations to update the CoreSet with.
}

\strong{dateCreated}: \code{character(1)} The date the \code{CoreSet} object was
created, as returned by the \code{date()} function.

\strong{dateCreated<-}: Update the 'dateCreated' item in the \code{annotation} slot of
a \code{CoreSet} object. Arguments:
\itemize{
\item value: A \code{character(1)} vector, as returned by the \code{date()} function.
}

\strong{cellInfo}: \code{data.frame} Metadata for all cell-lines in a \code{CoreSet} object.

\strong{cellInfo<-}: assign updated cell-line annotations to the \code{CoreSet} object. Arguments:

\strong{cellNames}: \code{character} Retrieve the rownames of the cellInfo slot from a \code{CoreSet} object.

\strong{cellNames<-}: assign new rownames to the cellInfo slot \code{data.frame} for a \code{CoreSet} object. Arguments:
\itemize{
\item value:
}

\strong{curation}: A \code{list} of curated mappings between identifiers in the
\code{CoreSet} object and the original data publication. Contains two
\code{data.frame}s, 'cell' with cell-line ids and 'tissue' with tissue ids.

\strong{curation<-}: Update the \code{curation} slot of a \code{CoreSet} object. Arugments:
\itemize{
\item value: A \code{list} of \code{data.frame}s, one for each type of curated identifier. For a \code{CoreSet} object the slot should contain tissue and cell-line id \code{data.frame}s.
}

\strong{datasetType}: \code{character(1)} The type treatment response in the
\code{sensitivity} slot. Valid values are 'sensitivity', 'perturbation' or 'both'.

\strong{datasetType<-}: Update the datasetType slot of a \code{CoreSet} object. Arguments:
\itemize{
\item value: A \code{character(1)} vector with
}
}
\examples{
dateCreated(clevelandSmall_cSet)

dateCreated(clevelandSmall_cSet)

cellInfo <- cellInfo(clevelandSmall_cSet)

cellNames(clevelandSmall_cSet)

cellNames(clevelandSmall_cSet) <- cellNames(clevelandSmall_cSet)

curation(clevelandSmall_cSet)

curation(clevelandSmall_cSet) <- curation(clevelandSmall_cSet)

datasetType(clevelandSmall_cSet)

datasetType(clevelandSmall_cSet) <- 'both'

}
