% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-accessors.R
\name{CoreSet-accessors}
\alias{CoreSet-accessors}
\alias{annotation,CoreSet-method}
\alias{annotation}
\alias{annotation<-,CoreSet,list-method}
\alias{annotation<-}
\alias{cellInfo,CoreSet-method}
\alias{cellInfo<-,CoreSet,data.frame-method}
\title{Accessing and modifying information in \code{CoreSet}}
\usage{
\S4method{annotation}{CoreSet}(object)

\S4method{annotation}{CoreSet,list}(object) <- value

\S4method{cellInfo}{CoreSet}(object)

\S4method{cellInfo}{CoreSet,data.frame}(object) <- value
}
\arguments{
\item{object}{A \code{CoreSet} object.}

\item{value}{See details.}

\item{...}{See details.}
}
\value{
Accessors:
\itemize{
\item cellInfo
}

Setters: An updated \code{CoreSet} object, returned invisibly.
}
\description{
Accessing and modifying information in \code{CoreSet}
}
\details{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{CoreSet} object.

annotation: getter method for annotation slot.

annotation<-: setter method for the annotation slot.
\itemize{
\item parameters:
\itemize{
\item value: a \code{list} of annotations to update the \code{CoreSet} with.
}
}



cellInfo<-: assign updated cell-line annotations to the `r .local_class` object
- parameters:
}
\seealso{
\code{\linkS4class{CoreSet}}
}
\concept{CoreSet-accessors}
