% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R, R/cellInfo-methods.R,
%   R/cellNames-methods.R, R/curation-methods.R, R/datasetType-methods.R,
%   R/dateCreated-methods.R, R/fNames-methods.R, R/featureInfo-methods.R,
%   R/mDataNames-methods.R, R/molecularProfiles-methods.R,
%   R/molecularProfilesSlot-methods.R, R/name-methods.R, R/pertNumber-methods.R,
%   R/phenoInfo-methods.R, R/sensNumber-methods.R, R/sensitivityInfo-methods.R,
%   R/sensitivityMeasures-methods.R, R/sensitivityProfiles-methods.R,
%   R/sensitivityRaw-methods.R, R/sensitivitySlot-methods.R
\docType{class}
\name{CoreSet-class}
\alias{CoreSet-class}
\alias{.CoreSet}
\alias{cellInfo,CoreSet-method}
\alias{cellInfo<-,CoreSet,data.frame-method}
\alias{cellNames,CoreSet-method}
\alias{cellNames<-,CoreSet,character-method}
\alias{curation,CoreSet-method}
\alias{curation<-,CoreSet,list-method}
\alias{datasetType,CoreSet-method}
\alias{datasetType<-,CoreSet-method}
\alias{dateCreated,CoreSet-method}
\alias{dateCreated<-,CoreSet-method}
\alias{fNames,CoreSet-method}
\alias{fNames<-,CoreSet,character,character-method}
\alias{featureInfo,CoreSet-method}
\alias{featureInfo<-,CoreSet,character,data.frame-method}
\alias{featureInfo<-,CoreSet,character,DataFrame-method}
\alias{mDataNames,CoreSet-method}
\alias{mDataNames<-,CoreSet-method}
\alias{molecularProfiles,CoreSet-method}
\alias{molecularProfiles<-,CoreSet,character,character,matrix-method}
\alias{molecularProfiles<-,CoreSet,character,missing,matrix-method}
\alias{molecularProfilesSlot,CoreSet-method}
\alias{molecularProfilesSlot<-,CoreSet,list-method}
\alias{name,CoreSet-method}
\alias{name<-,CoreSet-method}
\alias{pertNumber,CoreSet-method}
\alias{pertNumber<-,CoreSet,array-method}
\alias{phenoInfo,CoreSet-method}
\alias{phenoInfo<-,CoreSet,character,data.frame-method}
\alias{phenoInfo<-,CoreSet,character,DataFrame-method}
\alias{sensNumber,CoreSet-method}
\alias{sensNumber<-,CoreSet,matrix-method}
\alias{sensitivityInfo,CoreSet-method}
\alias{sensitivityInfo<-,CoreSet,data.frame-method}
\alias{sensitivityMeasures,CoreSet-method}
\alias{sensitivityMeasures<-,CoreSet,character-method}
\alias{sensitivityProfiles,CoreSet-method}
\alias{sensitivityProfiles<-,CoreSet,data.frame-method}
\alias{sensitivityProfiles<-,CoreSet,matrix-method}
\alias{sensitivityRaw,CoreSet-method}
\alias{sensitivityRaw<-,CoreSet,array-method}
\alias{sensitivitySlot,CoreSet-method}
\alias{sensitivitySlot<-,CoreSet,list-method}
\title{A Superclass to Contain Data for Genetic Profiling and Viability Screens of Cancer Cell Lines}
\usage{
\S4method{cellInfo}{CoreSet}(object)

\S4method{cellInfo}{CoreSet,data.frame}(object) <- value

\S4method{cellNames}{CoreSet}(object, ...)

\S4method{cellNames}{CoreSet,character}(object) <- value

\S4method{curation}{CoreSet}(object)

\S4method{curation}{CoreSet,list}(object) <- value

\S4method{datasetType}{CoreSet}(object)

\S4method{datasetType}{CoreSet}(object) <- value

\S4method{dateCreated}{CoreSet}(object)

\S4method{dateCreated}{CoreSet}(object) <- value

\S4method{fNames}{CoreSet}(object, mDataType)

\S4method{fNames}{CoreSet,character,character}(object, mDataType) <- value

\S4method{featureInfo}{CoreSet}(object, mDataType)

\S4method{featureInfo}{CoreSet,character,data.frame}(object, mDataType) <- value

\S4method{featureInfo}{CoreSet,character,DataFrame}(object, mDataType) <- value

\S4method{mDataNames}{CoreSet}(object)

\S4method{mDataNames}{CoreSet}(object) <- value

\S4method{molecularProfiles}{CoreSet}(object, mDataType, assay)

\S4method{molecularProfiles}{CoreSet,character,character,matrix}(object, mDataType, assay) <- value

\S4method{molecularProfiles}{CoreSet,character,missing,matrix}(object, mDataType, assay) <- value

\S4method{molecularProfilesSlot}{CoreSet}(object)

\S4method{molecularProfilesSlot}{CoreSet,list}(object) <- value

\S4method{name}{CoreSet}(object)

\S4method{name}{CoreSet}(object) <- value

\S4method{pertNumber}{CoreSet}(object)

\S4method{pertNumber}{CoreSet,array}(object) <- value

\S4method{phenoInfo}{CoreSet}(object, mDataType)

\S4method{phenoInfo}{CoreSet,character,data.frame}(object, mDataType) <- value

\S4method{phenoInfo}{CoreSet,character,DataFrame}(object, mDataType) <- value

\S4method{sensNumber}{CoreSet}(object)

\S4method{sensNumber}{CoreSet,matrix}(object) <- value

\S4method{sensitivityInfo}{CoreSet}(object)

\S4method{sensitivityInfo}{CoreSet,data.frame}(object) <- value

\S4method{sensitivityMeasures}{CoreSet}(object)

\S4method{sensitivityMeasures}{CoreSet,character}(object) <- value

\S4method{sensitivityProfiles}{CoreSet}(object)

\S4method{sensitivityProfiles}{CoreSet,data.frame}(object) <- value

\S4method{sensitivityProfiles}{CoreSet,matrix}(object) <- value

\S4method{sensitivityRaw}{CoreSet}(object)

\S4method{sensitivityRaw}{CoreSet,array}(object) <- value

\S4method{sensitivitySlot}{CoreSet}(object)

\S4method{sensitivitySlot}{CoreSet,list}(object) <- value
}
\arguments{
\item{object}{A \code{CoreSet}}

\item{value}{A \code{list} of curations for the cell and tissues types in the cSet object}

\item{...}{A \code{list} to allow definition of new parameters on this generic}

\item{mDataType}{A \code{character} with the type of molecular data to return/update}

\item{assay}{\code{character} Name of the desired assay; if excluded
defaults to first assay in the SummarizedExperiment for the given
mDataType. Use \code{assayNames(molecularProfiles(object, dataType)}
to check which assays are available for a given molecular datatype.}
}
\value{
An object of the CoreSet class

A \code{list} of unique cell and tissue identifiers to check validity
  of a cSet

A copy of the \code{RadioSet} with the updated curation slot
}
\description{
The CoreSet (CSet) class was developed as a superclass for pSets in the 
PharmacoGx and RadioGx packages to contain the data generated in screens 
of cancer cell lines for their genetic profile and sensitivities to therapy
(Pharmacological or Radiation). This class is meant to be a superclass which 
is contained within the PharmacoSet (pSet) and RadioSet (RSet) objects 
exported by PharmacoGx and RadioGx. The format of the data is similar for 
both pSets and rSets, allowing much of the code to be abstracted into 
the CoreSet super-class. However, the models involved with quantifying 
cellular response to Pharmacological and Radiation therapy are widely 
different, and extension of the cSet class allows the
packages to apply the correct model for the given data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{cellInfo}: Returns the annotations for all the cell lines tested on in the CoreSet

\item \code{cellInfo<-}: Update the cell line annotations

\item \code{cellNames}: Return the cell names used in the dataset

\item \code{cellNames<-}: Update the cell names used in the dataset

\item \code{curation}: Retrieve the curation slot form a cSet

\item \code{curation<-}: Update the curation slot of a cSet

\item \code{datasetType}: Update the dataset type of an rSet and return a copy of
the updated object

\item \code{datasetType<-}: Update the dataset type of an rSet and return a copy of
the updated object

\item \code{dateCreated}: Return the date the CoreSet was created

\item \code{dateCreated<-}: Update the date a cSet was created

\item \code{fNames}: Return the feature names used in the dataset

\item \code{fNames<-}: Update the feature names used in a molecular profile

\item \code{featureInfo}: Return the feature info for the given molecular data

\item \code{featureInfo<-}: Replace the gene info for the molecular data

\item \code{featureInfo<-}: Replace the gene info for the molecular data

\item \code{mDataNames}: Return the molecular data types available in a cSet object

\item \code{mDataNames<-}: Return the molecular data types available in a cSet object

\item \code{molecularProfiles}: Return the given type of molecular data from the CoreSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from the CoreSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from the CoreSet

\item \code{molecularProfilesSlot}: Return a list containing all molecularProfiles in the cSet

\item \code{molecularProfilesSlot<-}: Update the contents of the molecularProfiles slot in a
CoreSet and returns an update copy

\item \code{name}: Return the name of the CoreSet

\item \code{name<-}: Return the name of the CoreSet

\item \code{pertNumber}: Return the summary of available perturbation
experiments

\item \code{pertNumber<-}: Update the summary of available perturbation
experiments

\item \code{phenoInfo}: Return the experiment info from the given type of molecular data in CoreSet

\item \code{phenoInfo<-}: Update the given type of molecular data experiment info in the CoreSet

\item \code{phenoInfo<-}: Update the given type of molecular data experiment info in the CoreSet

\item \code{sensNumber}: Return the summary of available sensitivity
experiments

\item \code{sensNumber<-}: Update the summary of available sensitivity
experiments

\item \code{sensitivityInfo}: Return the drug dose sensitivity experiment info

\item \code{sensitivityInfo<-}: Update the sensitivity experiment info

\item \code{sensitivityMeasures}: Returns the available sensitivity profile
summaries, for example, whether there are IC50 values available

\item \code{sensitivityMeasures<-}: Updates the sensitivity measures in a cSet object and returns the updated object

\item \code{sensitivityProfiles}: Return the phenotypic data for the drug dose sensitivity

\item \code{sensitivityProfiles<-}: Update the phenotypic data for the drug dose
sensitivity

\item \code{sensitivityProfiles<-}: Update the phenotypic data for the drug dose
sensitivity

\item \code{sensitivityRaw}: Get the raw dose and vaibility data for a cSet

\item \code{sensitivityRaw<-}: Set the raw dose and viability data for a cSet and return
and updated copty

\item \code{sensitivitySlot}: Retrieve the contents of the sensitivity slot

\item \code{sensitivitySlot<-}: Set the raw dose and viability data for a cSet and return
and updated copty
}}

\section{Slots}{

\describe{
\item{\code{annotation}}{A \code{list} of annotation data about the CoreSet,
including the \code{$name} and the session information for how the object
was creating, detailing the exact versions of R and all the packages used}

\item{\code{molecularProfiles}}{A \code{list} containing \code{SummarizedExperiment}s 
type object for holding data for RNA, DNA, SNP and Copy Number Variation 
measurements respectively, with associated \code{rowData} and \code{colData} 
containing the row and column metadata}

\item{\code{cell}}{A \code{data.frame} containg the annotations for all the cell 
lines profiled in the data set, across all data types}

\item{\code{sensitivity}}{A \code{list} containing all the data for the sensitivity 
experiments, including \code{$info}, a \code{data.frame} containing the 
experimental info,\code{$raw} a 3D \code{array} containing raw data, 
\code{$profiles}, a \code{data.frame} containing sensitivity profiles 
statistics, and \code{$n}, a \code{data.frame} detailing the number of 
experiments for each cell-drug/radiationInfo pair}

\item{\code{perturbation}}{A \code{list} containting \code{$n}, a \code{data.frame} 
summarizing the available perturbation data,}

\item{\code{curation}}{A \code{list} containing mappings for
\code{cell}, \code{tissue} names  used in the data set to universal 
identifiers used between different CoreSet objects}

\item{\code{datasetType}}{A \code{character} string of 'sensitivity', 
'perturbation', or both detailing what type of data can be found in the 
CoreSet, for proper processing of the data}
}}

\examples{
data(clevelandSmall_cSet)
curation(clevelandSmall_cSet)

data(clevelandSmall_cSet)
curation(clevelandSmall_cSet) <- curation(clevelandSmall_cSet)

}
