% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{DataMapper-accessors}
\alias{DataMapper-accessors}
\alias{rawdata,DataMapper-method}
\alias{rawdata}
\alias{rawdata<-,DataMapper,ANY-method}
\alias{rawdata<-}
\title{Accessing and modifying data in a \code{DataMapper} object.}
\usage{
\S4method{rawdata}{DataMapper}(object)

\S4method{rawdata}{DataMapper,ANY}(object) <- value
}
\arguments{
\item{object}{A \code{DataMapper} object to get or set data from.}

\item{value}{A \code{list}-like object to assign to the rawdata slot. Should be
a \code{data.frame} or \code{data.table} with the current implementation.}
}
\value{
Accessors: See details

Setters: An update \code{DataMapper} object, returned invisibly.
}
\description{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{DataMapper} object.
}
\details{
\strong{rawdata}: Get the raw data slot from a \code{DataMapper} object. Returns
a list-like containing one or more raw data inputs to the
\code{DataMapper} object.

\strong{rawdata}: Set the raw data slot from a \code{DataMapper} object.
\strong{value}: The \code{list}-like object to set for the rawdata slot. Note: this
currently only supports \code{data.frame} or \code{data.table} objects.
}
\seealso{
Other DataMapper-accessors: 
\code{\link{rawdata<-,LongTableDataMapper,list-method}}
}
\concept{DataMapper-accessors}
