% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\name{DataMapper-accessors}
\alias{DataMapper-accessors}
\alias{rawdata,DataMapper-method}
\alias{rawdata}
\alias{rawdata<-,DataMapper,ANY-method}
\alias{rawdata<-,DataMapper-method}
\title{Accessing and modifying data in a \code{DataMapper} object.}
\usage{
\S4method{rawdata}{DataMapper}(object)

\S4method{rawdata}{DataMapper,ANY}(object) <- value
}
\arguments{
\item{object}{A \code{DataMapper} object to get or set data from.}
}
\value{
Accessors: See details

Setters: An update \code{DataMapper} object, returned invisibly.
}
\description{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{DataMapper} object.
}
\details{
\strong{rawdata}: Get the raw data slot from a \code{DataMapper} object. Returns
a list-like containing one or more raw data inputs to the
\code{DataMapper} object.

\strong{rawdata}: Set the raw data slot from a \code{DataMapper} object.
\strong{value}: The list-like object to set for the rawdata slot. Note: this
currently only supports data.frame or data.table objects.
}
\seealso{
Other DataMapper-accessors: 
\code{\link{LongTableDataMapper-accessors}}
}
\concept{DataMapper-accessors}
