% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateCreated-methods.R
\name{dateCreated}
\alias{dateCreated}
\title{dateCreated Generic}
\usage{
dateCreated(object, ...)
}
\arguments{
\item{object}{A \code{CoreSet}}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
The date the CoreSet was created
}
\description{
A generic for the dateCreated method
}
\examples{
dateCreated(clevelandSmall_cSet)

}
