% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{dateCreated}
\alias{dateCreated}
\title{dateCreated Generic}
\usage{
dateCreated(cSet)
}
\arguments{
\item{cSet}{A \code{CoreSet}}
}
\value{
The date the CoreSet was created
}
\description{
A generic for the dateCreated method
}
\examples{
dateCreated(Cleveland_small)

}
