% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSupportVec.R
\name{.GetSupportVec}
\alias{.GetSupportVec}
\title{.GetSupportVec}
\usage{
.GetSupportVec(x, output_length = 1001)
}
\arguments{
\item{x}{An input vector of dosages}

\item{output_length}{The length of the returned support vector}
}
\description{
.GetSupportVec
}
