% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiltities.R
\name{.meshEval}
\alias{.meshEval}
\title{meshEval}
\usage{
.meshEval(
  x,
  y,
  f,
  guess,
  lower_bounds,
  upper_bounds,
  density,
  n,
  scale,
  family,
  trunc
)
}
\value{
A \code{numeric} vector containing initial guesses for the 
  dose-response curve parameters
}
\description{
generate an initial guess for dose-response curve parameters by evaluating 
the residuals at different lattice points of the search space
}
\keyword{##FIXME::}
\keyword{PharmacoGx?}
\keyword{Why}
\keyword{different}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{this}
