% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiltities.R
\name{.residual}
\alias{.residual}
\title{TODO:: Write documentation}
\usage{
.residual(
  x,
  y,
  n,
  pars,
  f,
  scale = 0.07,
  family = c("normal", "Cauchy"),
  trunc = FALSE
)
}
\description{
TODO:: Write documentation
}
\keyword{internal}
