% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.verifyKeyIntegrity}
\alias{.verifyKeyIntegrity}
\title{Ensure that all rowID and colID keys are valid}
\usage{
.verifyKeyIntegrity(rowData, colData, assays)
}
\arguments{
\item{rowData}{A [`data.table`]}

\item{colData}{A [`data.table`]}

\item{assays}{A [`list`]}
}
\description{
Ensure that all rowID and colID keys are valid
}
\keyword{internal}
