% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{featureInfo}
\alias{featureInfo}
\title{featureInfo Generic}
\usage{
featureInfo(cSet, mDataType)
}
\arguments{
\item{cSet}{The \code{CoreSet} to retrieve feature annotations from}

\item{mDataType}{the type of molecular data}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for featureInfo method
}
\examples{
featureInfo(Cleveland_small, "rna")

}
