% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fNames-methods.R
\name{fNames<-}
\alias{fNames<-}
\title{fNames<- Generic}
\usage{
fNames(object, mDataType) <- value
}
\arguments{
\item{object}{The \code{CoreSet} to update}

\item{mDataType}{The molecular data type to update}

\item{value}{A \code{character} vector of the new cell names}
}
\value{
Updated \code{CoreSet}
}
\description{
A generic for the fNames replacement method
}
\examples{
data(clevelandSmall_cSet)
fNames(clevelandSmall_cSet, "rna") <- fNames(clevelandSmall_cSet, "rna")

}
