% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-guessMapping.R
\name{guessMapping,LongTableDataMapper-method}
\alias{guessMapping,LongTableDataMapper-method}
\title{Guess which columns in raw experiment data map to which dimensions.}
\usage{
\S4method{guessMapping}{LongTableDataMapper}(object, groups, subset)
}
\arguments{
\item{groups}{A \code{list} containing one or more vector of column names
to group-by. The function uses these to determine 1:1 mappings between
the combination of columns in each vector and unique values in the raw
data columns.}

\item{subset}{A \code{logical} vector indicating whether to to subset out mapped
columns after each grouping. Must be a single \code{TRUE} or \code{FALSE} or have
the same length as groups, indicating whether to subset out mapped columns
after each grouping.}

\item{mapper}{A \code{LongTableDataMapper} object.}
}
\description{
Checks for columns which are uniquely identified
}
