% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{idCols}
\alias{idCols}
\title{Generic to access the unique id columns in an S4 object used to}
\usage{
idCols(object, ...)
}
\arguments{
\item{object}{An [`S4`] object to get id columns from.}

\item{...}{Allow new arguments to this generic.}
}
\description{
Generic to access the unique id columns in an S4 object used to
}
