% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immutable-class.R
\name{is.immutable}
\alias{is.immutable}
\title{Check if an R object inherits from the immutable S3-class.}
\usage{
is.immutable(object)
}
\arguments{
\item{object}{Any R object to check if it inherits from the "immutable"
S3-class.}
}
\value{
\code{logical(1)} Does the object inherit from the "immutable" S3-class?
}
\description{
Check if an R object inherits from the immutable S3-class.
}
\examples{
immutable_list <- immutable(as.list(1:5))
is.immutable(immutable_list)

}
