% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-class.R
\docType{class}
\name{list_OR_LongTable-class}
\alias{list_OR_LongTable-class}
\title{Function to combine two LongTables into a single LongTable}
\arguments{
\item{x}{A \code{LongTable} object}

\item{y}{A \code{LongTable} object

A class union to allow multiple types in a CoreSet slot}
}
\description{
Function to combine two LongTables into a single LongTable
}
