% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R, R/methods-[[.R,
%   R/methods-assay.R, R/methods-assayNames.R, R/methods-assays.R,
%   R/methods-colData.R, R/methods-dim.R, R/methods-dimnames.R,
%   R/methods-getIntern.R, R/methods-rowData.R
\docType{class}
\name{LongTable-class}
\alias{LongTable-class}
\alias{.LongTable}
\alias{[[,LongTable,ANY,ANY-method}
\alias{assay,LongTable,ANY-method}
\alias{assay<-,LongTable,character-method}
\alias{assayNames,LongTable-method}
\alias{assays,LongTable-method}
\alias{assays<-,LongTable,list-method}
\alias{colData,LongTable-method}
\alias{colData<-,LongTable,ANY-method}
\alias{dim,LongTable-method}
\alias{colnames,LongTable-method}
\alias{rownames,LongTable-method}
\alias{dimnames,LongTable-method}
\alias{getIntern,LongTable,character-method}
\alias{rowData,LongTable-method}
\alias{rowData<-,LongTable-method}
\title{LongTable class definition}
\usage{
\S4method{[[}{LongTable,ANY,ANY}(x, i, withDimnames = TRUE, metadata = withDimnames, keys = !withDimnames)

\S4method{assay}{LongTable,ANY}(x, i, withDimnames = FALSE, metadata = withDimnames, key = !withDimnames)

\S4method{assay}{LongTable,character}(x, i) <- value

\S4method{assayNames}{LongTable}(x)

\S4method{assays}{LongTable}(x, withDimnames = FALSE, metadata = withDimnames, key = !withDimnames)

\S4method{assays}{LongTable,list}(x) <- value

\S4method{colData}{LongTable}(x, key = FALSE)

\S4method{colData}{LongTable,ANY}(x) <- value

\S4method{dim}{LongTable}(x)

\S4method{colnames}{LongTable}(x)

\S4method{rownames}{LongTable}(x)

\S4method{dimnames}{LongTable}(x)

\S4method{getIntern}{LongTable,character}(object, x, ...)

\S4method{rowData}{LongTable}(x, key = FALSE, use.names = FALSE)

\S4method{rowData}{LongTable}(x) <- value
}
\arguments{
\item{x}{A [`LongTable`] object to modify.}

\item{i}{[`integer`] or [`character`] vector containing the index or name
of the assay to update.}

\item{withDimnames}{[`logical`] Should the returned assays be joined to
the row and column identifiers (i.e., the pseudo dimnames of the object).}

\item{metadata}{[`logical`] Should row and column metadata also be joined
to the returned assays. This is useful for modifying assays before
reconstructing a new LongTable.}

\item{keys}{[`logical`] Should the row and column keys also be returned?
Defaults to !withDimnames.}

\item{key}{[`logical`] Should the rowKey column also be returned? Defaults
to FALSE.}

\item{value}{A [`data.table`] or [`data.frame`] to update}

\item{object}{[`LongTable`]}

\item{...}{[`pairlist`] Addtional arguments to get or mget inside of the
function.}

\item{use.names}{[`logical`] This parameter is just here to stop matching
the key argument to use.names from the rowData generic. It doesn't do
anything at this time and can be ignored.}
}
\value{
[`LongTable`] object containing the assay data from a treatment
  response experiment

[`LongTable`] With updated assays slot.

[`character`] Names of the assays contained in the `LongTable`.

A [`list`] of `data.table` objects, one per assay in the object.

A copy of the [`LongTable`] with the assays modified.

A [`data.table`] containing row identifiers and metadata.

A copy of the [`LongTable`] object with the `colData`
  slot updated.

[`numeric`] Vector of object dimensions.

[`character`] Vector of column names.

[`character`] Vector of row names.

[`list`] List with two character vectors, one for row and one for
    column names.

value of x if length(x) == 1 else named list of values for all
    symbols in x

A [`data.table`] containing rowID, row identifiers, and row metadata.

A copy of the [`LongTable`] object with the `rowData`
  slot updated.
}
\description{
Define a private constructor method to be used to build a
  `LongTable` object.

Select an assay from within a LongTable object.

This is used as an alternative to R attributes for storing structural
metadata of S4 objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[}: Get an assay from a LongTable object. This method
returns the row and column annotations by default to make assignment
and aggregate operations easiers.

\item \code{assay}: Retrieve an assay `data.table` object from the
`assays` slot of a `LongTable` object.

\item \code{assay<-}: Add or replace an assay in a LongTable by name. Currently
this function only works when the assay has all columns in row and column
data tables (i.e., when assays is retured withDimnames=TRUE). This will
be fixed in future updates.

\item \code{assayNames}: Return the names of the assays contained in a
`LongTable`

\item \code{assays}: Get a list containing all the assays in a `LongTable`.

\item \code{assays<-}: Update the assays in a LongTable object. The rowIDs
and colIDs must be present in all assays to allow successfully remapping
the keys. We recommend modifying the list returned by
assays(longTable, withDimnames=TRUE) and the reassigning to the `LongTable`.

\item \code{colData}: Get the column level annotations for a LongTable
object.

\item \code{colData<-}: Upadte the colData of a LongTable object. Currently
requires that all of the colIDs(longTable) be in the value object.

\item \code{dim}: Get the number of row annotations by the number of
column annotations from a LongTable object. Please note that row x columns
does not necessarily equal the number of rows in an assay, since it is
not required for each assay to have every row or column present.

\item \code{colnames}: Retrieve the pseudo-colnames of a LongTable object,
these are constructed by pasting together the colIDs(longTable) and
can be used in the subset method for regex based queries.

\item \code{rownames}: Retrieve the pseudo-rownames of a LongTable object,
these are constructed by pasting together the rowIDs(longTable) and
can be used in the subset method for regex based queries.

\item \code{dimnames}: Get the pseudo-dimnames for a LongTable object. See
colnames and rownames for more information.

\item \code{getIntern}: Access structural metadata present within a
LongTable object. This is mostly for developmer use.

\item \code{rowData}: Get the row level annotations for a `LongTable` object.

\item \code{rowData<-}: Update the row annotations for a `LongTable` object.
Currently requires that all columns in rowIDs(longTable) be present in
value.
}}

\section{Slots}{

\describe{
\item{\code{rowData}}{A [`data.table`] containing the metadata associated with the
row dimension of a `LongTable`.}

\item{\code{colData}}{A [`data.table`] containing the metadata associated with the
column dimension of a `LongTable`.}

\item{\code{assays}}{A [`list`] of [`data.table`s], one for each assay in a
`LongTable`.}

\item{\code{metadata}}{An optional [`list`] of additional metadata for a `LongTable`
which doesn't map to one of the dimensions.}

\item{\code{.intern}}{An [`enivornment`] that holds internal structural metadata
about a LongTable object, such as which columns are required to key
the object. An environment has been used to allow locking items, which
can prevent accidental modification of a property required for the class
to work.}
}}

\examples{
merckLongTable[['viability']]

# Default annotations, just the key columns
assay(merckLongTable, 'viability')
assay(merckLongTable, 1)

# With identifiers joined
assay(merckLongTable, 'viability', withDimnames=TRUE)

# With identifiers and metadata
assay(merckLongTable, 'viability_summary', withDimnames=TRUE, metadata=TRUE)

assay(merckLongTable, 'viability') <- assay(merckLongTable, 'viability', withDimnames=TRUE)
assay(merckLongTable, 'viability') <- merckLongTable$viability

assayNames(merckLongTable)

assays(merckLongTable)

assays(merckLongTable) <- assays(merckLongTable, withDimnames=TRUE)

colData(merckLongTable)

# Get the keys as well, mostly for internal use
colData(merckLongTable, key=TRUE)

colData(merckLongTable) <- colData(merckLongTable)

dim(merckLongTable)

dim(merckLongTable)

colnames(merckLongTable)

rownames(merckLongTable)

dimnames(merckLongTable)

getIntern(merckLongTable, 'rowIDs')
getIntern(merckLongTable, c('colIDs', 'colMeta'))

rowData(merckLongTable)

rowData(merckLongTable) <- rowData(merckLongTable)

}
\keyword{internal}
