% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-class.R
\name{LongTable}
\alias{LongTable}
\title{LongTable constructor method}
\usage{
LongTable(
  rowData,
  rowIDs,
  colData,
  colIDs,
  assays,
  assayMap,
  metadata = list(),
  keep.rownames = FALSE
)
}
\value{
A \code{LongTable} object containing the data for a treatment response
experiment and configured according to the rowIDs and colIDs arguments.
}
\description{
LongTable constructor method
}
