% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R,
%   R/LongTableDataMapper-class.R
\name{LongTableDataMapper-accessors}
\alias{LongTableDataMapper-accessors}
\alias{rawdata,LongTableDataMapper-method}
\alias{rawdata}
\alias{rowDataMap}
\title{Accessing and modifying data in a \code{LongTableDataMapper} object.}
\usage{
rowDataMap(object, ...)

rowDataMap(object, ...)
}
\arguments{
\item{object}{An \code{S4} object to access the rowDataMap from.}

\item{...}{Allow new parameters to be defined for this generic.}
}
\value{
Accessors: See details

Setters: An update \code{LongTableDataMapper} object, returned invisibly.

\code{list} of two \code{character} vectors, the first are the columns required
to uniquely identify each row of a \code{LongTable} and the second any additional
row-level metadata.

Accessors: See details

Setters: An update \code{LongTableDataMapper} object, returned invisibly.

\code{list} of two \code{character} vectors, the first are the columns required
to uniquely identify each row of a \code{LongTable} and the second any additional
row-level metadata.
}
\description{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{LongTableDataMapper} object.

Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{LongTableDataMapper} object.
}
\details{
\strong{rawdata}: Get the raw data slot from a \code{LongTableDataMapper} object. Returns
a list-like containing one or more raw data inputs to the
\code{LongTableDataMapper} object.

\strong{rawdata}: Get the raw data slot from a \code{LongTableDataMapper} object. Returns
a list-like containing one or more raw data inputs to the
\code{LongTableDataMapper} object.
}
