% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMapper-class.R
\docType{class}
\name{LongTableDataMapper-class}
\alias{LongTableDataMapper-class}
\alias{.LongTableDataMapper}
\title{A Class for Mapping Between Raw Data and an \code{LongTable} Object}
\description{
A Class for Mapping Between Raw Data and an \code{LongTable} Object
}
\section{Slots}{

\itemize{
\item rowDataMap: A list of mappings from.
\item colDataMap: A list of mappings from.
\item assayMap: A list of mappings from .
\item metadataMap: A list of mappings from.
\item metadata: A list of mappings from.
}


\itemize{
\item rawdata: A list-like object containing one or more pieces of raw data
that will be processed and mapped to the slots of an \code{S4} object.
}
}

\keyword{internal}
