% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{mDataNames}
\alias{mDataNames}
\title{mDataNames Generic}
\usage{
mDataNames(cSet)
}
\arguments{
\item{cSet}{CoreSet object}
}
\value{
Vector of names of the molecular data types
}
\description{
A generic for the mDataNames method
}
\examples{
mDataNames(Cleveland_small)

}
