% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metaConstruct.R
\name{metaConstruct,LongTableDataMapper-method}
\alias{metaConstruct,LongTableDataMapper-method}
\title{metaConstruct}
\usage{
\S4method{metaConstruct}{LongTableDataMapper}(mapper)
}
\arguments{
\item{mapper}{An \code{LongTableDataMapper} object abstracting arguments to an
the \code{LongTable} constructor.}
}
\value{
A \code{LongTable} object, as specified in the mapper.
}
\description{
metaConstruct
}
