% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-molecularProfilesSlot.R
\name{molecularProfilesSlot<-}
\alias{molecularProfilesSlot<-}
\title{molecularProfilesSlot<-}
\usage{
molecularProfilesSlot(object) <- value
}
\arguments{
\item{object}{A \code{CoreSet} object for which values will be replaced}

\item{value}{A \code{list} containing molecular profiles as SummarizedExperiments}
}
\value{
A copy of the \code{CoreSet} with the molecularProfiles slot updated
}
\description{
Replace method for the molecular profiles slot of a cSet
}
\examples{
data(clevelandSmall_cSet)
molecularProfilesSlot(clevelandSmall_cSet) <- molecularProfilesSlot(clevelandSmall_cSet)

}
