% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name-methods.R
\name{name}
\alias{name}
\title{name Generic}
\usage{
name(object, ...)
}
\arguments{
\item{object}{A \code{CoreSet}}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
The name of the CoreSet
}
\description{
A generic for the name method
}
\examples{
name(clevelandSmall_cSet)

}
