% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immutable-class.R
\name{print.immutable}
\alias{print.immutable}
\title{Print method for objects inheriting from the "immutable" S3-class}
\usage{
\method{print}{immutable}(x)
}
\arguments{
\item{x}{An object inheriting from the "immutable" S3-class.}
}
\value{
None, \code{invisible(NULL)}
}
\description{
Print method for objects inheriting from the "immutable" S3-class
}
