% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{sensitivityInfo<-}
\alias{sensitivityInfo<-}
\title{sensitivityInfo<- Generic}
\usage{
sensitivityInfo(object) <- value
}
\arguments{
\item{object}{The \code{CoreSet} to update}

\item{value}{A \code{data.frame} with the new sensitivity annotations}
}
\value{
Updated \code{CoreSet}
}
\description{
A generic for the sensitivityInfo replacement method
}
\examples{
sensitivityInfo(Cleveland_small) <- sensitivityInfo(Cleveland_small)

}
