% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityInfo}
\alias{sensitivityInfo}
\title{Generic function to get the annotations for a treatment response experiment
  from an S4 class}
\usage{
sensitivityInfo(object, ...)
}
\arguments{
\item{object}{An [`S4`] object to get treatment response experiment
annotations from.}

\item{...}{Allow new arguments to be defined for this generic.}
}
\value{
Depends on the implemented method
}
\description{
Generic function to get the annotations for a treatment response experiment
  from an S4 class
}
\examples{
print("Generics shouldn't need examples?")

}
