% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{sensitivityInfo}
\alias{sensitivityInfo}
\title{sensitivityInfo Generic}
\usage{
sensitivityInfo(cSet)
}
\arguments{
\item{cSet}{The \code{CoreSet} to retrieve sensitivity experiment annotations from}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for sensitivityInfo method
}
\examples{
sensitivityInfo(Cleveland_small)

}
