% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityMeasures-methods.R
\name{sensitivityMeasures<-}
\alias{sensitivityMeasures<-}
\title{sensitivityMeasures<- Generic}
\usage{
sensitivityMeasures(object, ...) <- value
}
\arguments{
\item{object}{The \code{CoreSet} object to update}

\item{...}{Fallthrough arguements for defining new methods}

\item{value}{A \code{character} vector of sensitivity measures to use to update the cSet object}
}
\value{
A update \code{CoreSet} object with the new sensitivity measures
}
\description{
A generic for the sensitivityMeasure<- method
}
\examples{
sensitivityMeasures(clevelandSmall_cSet) <- sensitivityMeasures(clevelandSmall_cSet)

}
