% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityMeasures}
\alias{sensitivityMeasures}
\title{sensitivityMeasures Generic}
\usage{
sensitivityMeasures(object, ...)
}
\arguments{
\item{object}{An [`S4`] object to retrieve the names of sensitivty summary
measurements for.}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
Depends on the implemented method
}
\description{
Get the names of the sensitivity summary metrics available in an S4
  object.
}
\examples{
sensitivityMeasures(clevelandSmall_cSet)

}
