% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{sensitivityMeasures}
\alias{sensitivityMeasures}
\title{sensitivityMeasures Generic}
\usage{
sensitivityMeasures(cSet)
}
\arguments{
\item{cSet}{The \code{CoreSet}}
}
\value{
A \code{character} vector of all the available sensitivity measures
}
\description{
A generic for the sensitivityMeasures  method
}
\examples{
sensitivityMeasures(Cleveland_small)

}
