% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityProfiles-methods.R
\name{sensitivityProfiles}
\alias{sensitivityProfiles}
\title{sensitivityProfiles Generic}
\usage{
sensitivityProfiles(object, ...)
}
\arguments{
\item{object}{The \code{CoreSet} to retrieve sensitivity experiment data from}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for sensitivityProfiles method
}
\examples{
sensitivityProfiles(clevelandSmall_cSet)

}
