% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivityProfiles}
\alias{sensitivityProfiles}
\title{sensitivityProfiles Generic}
\usage{
sensitivityProfiles(object, ...)
}
\arguments{
\item{object}{The \code{S4} object to retrieve sensitivity profile summaries
from.}

\item{...}{\code{pairlist} Allow defining new arguments for this generic.}
}
\value{
Depends on the implemented method
}
\description{
A generic for sensitivityProfiles getter method
}
\examples{
print("Generics shouldn't need examples?")

}
