% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{sensitivityProfiles}
\alias{sensitivityProfiles}
\title{sensitivityProfiles Generic}
\usage{
sensitivityProfiles(cSet)
}
\arguments{
\item{cSet}{The \code{CoreSet} to retrieve sensitivity experiment data from}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for sensitivityProfiles method
}
\examples{
sensitivityProfiles(Cleveland_small)

}
