% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityRaw-methods.R
\name{sensitivityRaw<-}
\alias{sensitivityRaw<-}
\title{sensitivityRaw<- Generic}
\usage{
sensitivityRaw(object, ...) <- value
}
\arguments{
\item{object}{An \code{object} to extract the raw sensitivity data from}

\item{...}{A \code{list} to allow new parameters in specific methods}

\item{value}{A 3D \code{array} containing dose and viability metrics to update the object with}
}
\value{
A \code{array} containing the raw sensitivity data as experiment by dose level by metric
}
\description{
#' @examples
data(clevelandSmall_cSet)
sensitivityRaw(clevelandSmall_cSet) <- sensitivityRaw(clevelandSmall_cSet)
}
