% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityRaw-methods.R
\name{sensitivityRaw}
\alias{sensitivityRaw}
\title{sensitivityRaw Generic}
\usage{
sensitivityRaw(object, ...)
}
\arguments{
\item{object}{An \code{object} to extract the raw sensitivity data from}

\item{...}{A \code{list} to allow new parameters in specific methods}
}
\value{
A \code{array} containing the raw sensitivity data as experiment by dose level by metric
}
\description{
sensitivityRaw Generic
}
\examples{
data(clevelandSmall_cSet)
sensRaw <- sensitivityRaw(clevelandSmall_cSet)
head(sensRaw)

}
