% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{sensitivitySlotToLongTable}
\alias{sensitivitySlotToLongTable}
\title{sensitivitySlotToLongTable Generic}
\usage{
sensitivitySlotToLongTable(object, ...)
}
\arguments{
\item{object}{[`CoreSet`] Object inheriting from CoreSet.}

\item{...}{Allow new arguments to be defined on this generic.}
}
\value{
A [`LongTable`] object containing the data in the sensitivity slot.
}
\description{
Convert the sensitivity slot in an object inheriting from a CoreSet from a
  list to a LongTable.
}
\examples{
print("Generics shouldn't need examples?")

}
