% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSet-accessors.R
\name{sensNumber}
\alias{sensNumber}
\title{sensNumber Generic}
\usage{
sensNumber(object, ...)
}
\arguments{
\item{object}{A \code{CoreSet}}

\item{...}{Fallthrough arguements for defining new methods}
}
\value{
A \code{data.frame} with the number of sensitivity experiments per drug and cell line
}
\description{
A generic for the sensNumber method
}
\examples{
sensNumber(clevelandSmall_cSet)

}
