% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{sensNumber}
\alias{sensNumber}
\title{sensNumber Generic}
\usage{
sensNumber(cSet)
}
\arguments{
\item{cSet}{A \code{CoreSet}}
}
\value{
A \code{data.frame} with the number of sensitivity experiments per drug and cell line
}
\description{
A generic for the sensNumber method
}
\examples{
sensNumber(Cleveland_small)

}
