% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{showSigAnnot}
\alias{showSigAnnot}
\title{Get the annotations for a `Signature` class object, as returned by `drugSensitivitysig` or `radSensitivtySig` functions
   available in `PharmacoGx` and `RadioGx`, respectively.}
\usage{
showSigAnnot(object, ...)
}
\arguments{
\item{object}{A `Signature` class object}

\item{...}{Allow definition of new arguments to this generic}
}
\value{
NULL Prints the signature annotations to console
}
\description{
Get the annotations for a `Signature` class object, as returned by `drugSensitivitysig` or `radSensitivtySig` functions
   available in `PharmacoGx` and `RadioGx`, respectively.
}
