% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongTable-utils.R
\name{subset,LongTable-method}
\alias{subset,LongTable-method}
\title{Subset method for a LongTable object.}
\usage{
\S4method{subset}{LongTable}(x, i, j, assays, reindex = TRUE)
}
\arguments{
\item{x}{\code{LongTable} The object to subset.}

\item{i}{\code{character}, \code{numeric}, \code{logical} or \code{expression}
Character: pass in a character vector of drug names, which will subset the
object on all row id columns matching the vector.
Numeric or Logical: these select based on the rowKey from the \code{rowData}
method for the \code{LongTable}.
Call: Accepts valid query statements to the \code{data.table} i parameter,
this can be used to make complex queries using the \code{data.table} API
for the \code{rowData} data.table.}

\item{j}{\code{character}, \code{numeric}, \code{logical} or \code{expression}
Character: pass in a character vector of drug names, which will subset the
object on all drug id columns matching the vector.
Numeric or Logical: these select based on the rowID from the \code{rowData}
method for the \code{LongTable}.
Call: Accepts valid query statements to the \code{data.table} i parameter,
this can be used to make complex queries using the \code{data.table} API
for the \code{colData} data.table.}

\item{assays}{\code{character}, \code{numeric} or \code{logical} Optional list of assay
names to subset. Can be used to subset the assays list further,
returning only the selected items in the new LongTable.}

\item{reindex}{\code{logical} Should the col/rowKeys be remapped after subsetting.
defaults to TRUE. For chained subsetting you may be able to get performance
gains by setting to FALSE and calling reindex() manually after subsetting
is finished.}
}
\value{
\code{LongTable} A new \code{LongTable} object subset based on the specified
parameters.
}
\description{
Allows use of the colData and rowData \code{data.table} objects to query based on
rowID and colID, which is then used to subset all value data.tables stored
in the dataList slot.
This function is endomorphic, it always returns a LongTable object.
}
\examples{
# Character
subset(merckLongTable, 'ABT-888', 'CAOV3')
# Numeric
subset(merckLongTable, 1, c(1, 2))
# Logical
subset(merckLongTable, , colData(merckLongTable)$sampleid == 'A2058')
# Call
subset(merckLongTable, drug1id == 'Dasatinib' & drug2id != '5-FU',
    sampleid == 'A2058')

}
