% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subsetTo}
\alias{subsetTo}
\title{Subset a CoreSet object based on various parameters, such as cell lines, molecular features}
\usage{
subsetTo(object, ...)
}
\arguments{
\item{object}{An object inheriting from the `CoreGx::CoreSet` class}

\item{...}{Allow definition of new arguments to this generic}
}
\value{
A subsetted version of the original `object`
}
\description{
Subset a CoreSet object based on various parameters, such as cell lines, molecular features
}
