% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summarizeSensitivityProfiles}
\alias{summarizeSensitivityProfiles}
\title{Summarize across replicates for a sensitivity dose-response experiment}
\usage{
summarizeSensitivityProfiles(object, ...)
}
\arguments{
\item{object}{An object inheriting form the `CoreGx::CoreSet` class}

\item{...}{Allow definition of new arguments to this generic}
}
\value{
[`data.frame`] containing treatment by cell line summary of a sensitivity experiment with values as the selected `sensitivity.measure`. Defaults `sensitivity.measure` is `auc_recomputed`.
}
\description{
Summarize across replicates for a sensitivity dose-response experiment
}
\examples{
# Dummy examples to satisfy Bioc requiremnts
setMethod('summarizeSensitivityProfiles', signature('CoreSet'), function(object) object)

}
\keyword{internal}
