% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSetClass.R
\name{updateCellId}
\alias{updateCellId}
\title{Update the cell ids in a cSet object}
\usage{
updateCellId(object, new.ids = vector("character"))
}
\arguments{
\item{object}{The object for witch the cell ids will be updated}

\item{new.ids}{The new ids to assign to the object}
}
\description{
Update the cell ids in a cSet object
}
\keyword{internal}
